% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfana.R, R/wrassp_re-export.R
\name{acfana}
\alias{acfana}
\title{acfana}
\usage{
acfana(
  listOfFiles = NULL,
  optLogFilePath = NULL,
  beginTime = 0,
  centerTime = FALSE,
  endTime = 0,
  windowShift = 5,
  windowSize = 20,
  effectiveLength = TRUE,
  window = "BLACKMAN",
  analysisOrder = 0,
  energyNormalization = FALSE,
  lengthNormalization = FALSE,
  toFile = TRUE,
  explicitExt = NULL,
  outputDirectory = NULL,
  forceToLog = useWrasspLogger,
  verbose = TRUE
)

acfana(
  listOfFiles = NULL,
  optLogFilePath = NULL,
  beginTime = 0,
  centerTime = FALSE,
  endTime = 0,
  windowShift = 5,
  windowSize = 20,
  effectiveLength = TRUE,
  window = "BLACKMAN",
  analysisOrder = 0,
  energyNormalization = FALSE,
  lengthNormalization = FALSE,
  toFile = TRUE,
  explicitExt = NULL,
  outputDirectory = NULL,
  forceToLog = useWrasspLogger,
  verbose = TRUE
)
}
\arguments{
\item{listOfFiles}{vector of file paths to be processed by function}

\item{optLogFilePath}{path to option log file}

\item{beginTime}{= \if{html}{\out{<time>}}: set begin of analysis interval to \if{html}{\out{<time>}} seconds (default: 0 = beginning of file)}

\item{centerTime}{= \if{html}{\out{<time>}}: set single-frame analysis with the analysis window centred at \if{html}{\out{<time>}} seconds;
overrules BeginTime, EndTime and WindowShift options}

\item{endTime}{= \if{html}{\out{<time>}}: set end of analysis interval to \if{html}{\out{<time>}} seconds (default: 0 = end of file)}

\item{windowShift}{= \if{html}{\out{<dur>}}: set analysis window shift to \if{html}{\out{<dur>}} ms (default: 5.0)}

\item{windowSize}{= \if{html}{\out{<dur>}}: set analysis window size to \if{html}{\out{<dur>}} ms; overrules EffectiveLength parameter}

\item{effectiveLength}{make window size effective rather than exact}

\item{window}{= \if{html}{\out{<type>}}: set analysis window function to \if{html}{\out{<type>}} (default: BLACKMAN)}

\item{analysisOrder}{= \if{html}{\out{<num>}}: set analysis order to \if{html}{\out{<num>}} (default: 0 = sample rate in kHz + 3)}

\item{energyNormalization}{calculate energy-normalized autocorrelation}

\item{lengthNormalization}{calculate length-normalized autocorrelation}

\item{toFile}{write results to file (default extension is .acf)}

\item{explicitExt}{set if you wish to override the default extension}

\item{outputDirectory}{directory in which output files are stored. Defaults to NULL, i.e.
the directory of the input files}

\item{forceToLog}{is set by the global package variable useWrasspLogger. This is set
to FALSE by default and should be set to TRUE is logging is desired.}

\item{verbose}{display infos & show progress bar}
}
\value{
nrOfProcessedFiles or if only one file to process return AsspDataObj of that file
}
\description{
acfana function adapted from libassp
}
\details{
Analysis of short-term autocorrelation function of
the signals in \if{html}{\out{<listOFFiles>}}.
Analysis results will be written to a file with the
base name of the input file and extension '.acf'.
Default output is in SSFF binary format (track 'acf').
}
\examples{
# get path to audio file
path2wav <- list.files(system.file("extdata", package = "wrassp"), 
                       pattern = glob2rx("*.wav"), 
                       full.names = TRUE)[1]

# calculate short-term autocorrelation
res <- acfana(path2wav, toFile=FALSE)

# plot short-term autocorrelation values
matplot(seq(0,numRecs.AsspDataObj(res) - 1) / rate.AsspDataObj(res) + 
        attr(res, 'startTime'), 
        res$acf, 
        type='l', 
        xlab='time (s)', 
        ylab='short-term autocorrelation values')
        
}
\seealso{
\link[wrassp:acfana]{wrassp::acfana}
}
\author{
Raphael Winkelmann

Lasse Bombien
}
