% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_python_pm_pvoice_tremor.R
\name{analyze_amplitude_tremor_pladdrr}
\alias{analyze_amplitude_tremor_pladdrr}
\title{Analyze amplitude tremor using pladdrr}
\usage{
analyze_amplitude_tremor_pladdrr(
  sound,
  slength,
  analysisTimeStep,
  minPitch,
  silenceThreshold,
  voicingThreshold,
  octaveCost,
  octaveJumpCost,
  voicedUnvoicedCost,
  minTremorFreq,
  maxTremorFreq,
  tremorMagThresh,
  tremorCyclicalThresh,
  ampTremorOctaveCost,
  nanAsZero
)
}
\description{
Analyze amplitude tremor using pladdrr
}
\keyword{internal}
