% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_python_pm_pvoice_tremor.R
\name{analyze_frequency_tremor_pladdrr}
\alias{analyze_frequency_tremor_pladdrr}
\title{Analyze frequency tremor using pladdrr}
\usage{
analyze_frequency_tremor_pladdrr(
  sound,
  slength,
  analysisTimeStep,
  minPitch,
  maxPitch,
  silenceThreshold,
  voicingThreshold,
  octaveCost,
  octaveJumpCost,
  voicedUnvoicedCost,
  minTremorFreq,
  maxTremorFreq,
  tremorMagThresh,
  tremorCyclicalThresh,
  freqTremorOctaveCost,
  nanAsZero
)
}
\description{
Analyze frequency tremor using pladdrr
}
\keyword{internal}
