% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_pharyngeal.R
\name{analyze_pharyngeal_mid}
\alias{analyze_pharyngeal_mid}
\title{Internal: Analyze Mid}
\usage{
analyze_pharyngeal_mid(
  sound,
  sampfreq,
  start_time,
  end_time,
  f0_mid,
  f1_mid,
  f2_mid,
  f3_mid,
  bw1_mid_norm,
  bw2_mid_norm,
  bw3_mid_norm,
  UNDEFINED
)
}
\description{
Internal: Analyze Mid
}
\keyword{internal}
