% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_pharyngeal.R
\name{analyze_pharyngeal_onset}
\alias{analyze_pharyngeal_onset}
\title{Internal: Analyze Onset}
\usage{
analyze_pharyngeal_onset(
  sound,
  sampfreq,
  start_time,
  end_time,
  f0_start,
  f1_start,
  f2_start,
  f3_start,
  bw1_start_norm,
  bw2_start_norm,
  bw3_start_norm,
  UNDEFINED
)
}
\description{
Internal: Analyze Onset
}
\keyword{internal}
