% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_ssff.R
\name{aperiodicities}
\alias{aperiodicities}
\title{Estimates aperiodicity of a speech signal}
\usage{
aperiodicities(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 70,
  maxF = 200,
  voiced_voiceless_threshold = 0.01,
  explicitExt = "wap",
  outputDirectory = NULL,
  toFile = TRUE
)
}
\arguments{
\item{listOfFiles}{Vector of file paths to process}

\item{beginTime}{Start time in seconds (default: 0.0)}

\item{endTime}{End time in seconds (default: 0.0 = end of file)}

\item{windowShift}{Frame shift in milliseconds (default: 10.0)}

\item{minF}{Minimum F0 in Hz (default: 60.0)}

\item{maxF}{Maximum F0 in Hz (default: 400.0)}

\item{explicitExt}{Output file extension (default: "f0")}

\item{outputDirectory}{Output directory (default: NULL = same as input)}

\item{toFile}{Write results to file (default: TRUE)}
}
\value{
An SSFF track object containing two tracks (f0 and corr) that are
either returned (toFile == FALSE) or stored on disk.
}
\description{
Aperiodicity is estimated using function for assessing band-aperiodicities
using the Definitive Decomposition Derived Dirt-Cheap (D4C) algorithm
\insertCite{Morise.2016.10.1016/j.specom.2016.09.001}{superassp} implemented
in the WORLD vocoder
\insertCite{MORISE.2016.10.1587/transinf.2015edp7457}{superassp}. The the \link[=DIO]{dio}
\insertCite{morise2010rapid}{superassp} pitch algorithm is used to calculate
the periodic component.
}
\references{
\insertAllCited{}
}
