% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_avaudio.R
\name{avaudio_to_av}
\alias{avaudio_to_av}
\title{Convert AVAudio to av::read_audio_bin Format}
\usage{
avaudio_to_av(audio)
}
\arguments{
\item{audio}{AVAudio object}
}
\value{
Integer vector with attributes (channels, sample_rate)
}
\description{
Convert AVAudio object back to the format returned by av::read_audio_bin(),
which is an integer vector with channels and sample_rate attributes.
}
\examples{
\dontrun{
audio <- read_avaudio("speech.wav")
audio_vec <- avaudio_to_av(audio)

# audio_vec is now compatible with av::read_audio_bin output
}

}
