% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat_internal_iaif.R
\name{calc_residual_r}
\alias{calc_residual_r}
\title{Per-GCI LPC analysis and inverse filtering
R wrapper around the Rcpp implementation of \code{calc_residual.m}.
If GCI vector is too short, falls back to pure R.}
\usage{
calc_residual_r(x, x_lpc, ord_lpc, GCI)
}
\description{
Per-GCI LPC analysis and inverse filtering
R wrapper around the Rcpp implementation of \code{calc_residual.m}.
If GCI vector is too short, falls back to pure R.
}
\keyword{internal}
