% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_pdsi.R
\name{calculate_jitter_ppq5_pladdrr}
\alias{calculate_jitter_ppq5_pladdrr}
\title{Calculate Jitter PPQ5 from Stable Vowel}
\usage{
calculate_jitter_ppq5_pladdrr(df)
}
\description{
Calculate Jitter PPQ5 from Stable Vowel
}
\keyword{internal}
