% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_dsi.R
\name{calculate_mpt_pladdrr}
\alias{calculate_mpt_pladdrr}
\title{Calculate Maximum Phonation Time}
\usage{
calculate_mpt_pladdrr(df)
}
\description{
Calculate Maximum Phonation Time
}
\keyword{internal}
