% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice_sauce.R
\name{check_voice_sauce_status}
\alias{check_voice_sauce_status}
\title{Check VoiceSauce Status on Package Load}
\usage{
check_voice_sauce_status()
}
\value{
Invisible NULL
}
\description{
Internal function called during package startup to check VoiceSauce status.
}
\keyword{internal}
