% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_c_assp_rmsana.R, R/superassp_fileHelper.R
\name{cleanupConvertedInputMediaFiles}
\alias{cleanupConvertedInputMediaFiles}
\title{Cleanup converted input media files}
\usage{
cleanupConvertedInputMediaFiles(toClear, keepConverted, verbose)

cleanupConvertedInputMediaFiles(toClear, keepConverted, verbose)
}
\arguments{
\item{toClear}{Character vector of files to remove}

\item{keepConverted}{Logical whether to keep converted files}

\item{verbose}{Logical whether to show progress messages}
}
\description{
Cleanup converted input media files

Cleanup converted input media files
}
\keyword{internal}
