% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat.R
\name{clear_dsp_environment}
\alias{clear_dsp_environment}
\title{Tear down a DSP environment}
\usage{
clear_dsp_environment(dsp_environment_path)
}
\arguments{
\item{dsp_environment_path}{The full path to a directory set up for storing sound files
for processing.}
}
\value{
The function returns \code{FALSE} if the function was \emph{unable} to remove
the environment. If the function returns \code{TRUE}, the environment associated
with the speech signal file was successfully removed, or did not exist in
the first place.This means that the user can be reasonably sure in this
case to not have clashing environments for Praat to run in if this function returns
\code{TRUE}.
}
\description{
This function will remove the environment that has been created to process a
specific speech signal file \code{signalFileFullPath}. The same signal file path
should be given to this function and the \link{make_dsp_environment} function to
ensure that the correct environment is removed.
}
