% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat_internal_creak.R
\name{compute_delta}
\alias{compute_delta}
\title{Compute delta (first derivative) of feature matrix}
\usage{
compute_delta(feat)
}
\description{
Compute delta (first derivative) of feature matrix
}
\keyword{internal}
