% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxit_analysis_exports.R
\name{compute_intensity_mean_cpp}
\alias{compute_intensity_mean_cpp}
\title{Intensity mean}
\usage{
compute_intensity_mean_cpp(lin_power)
}
\arguments{
\item{lin_power}{Linear power vector (0–1)}
}
\value{
Mean intensity in dB
}
\description{
Convert linear power to dB and compute mean.
}
