% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\alias{PHONET_SAMPLE_RATE}
\alias{DEFAULT_CHANNELS}
\alias{SACC_SAMPLE_RATE}
\alias{BROUHAHA_SAMPLE_RATE}
\alias{MS_TO_SECONDS}
\alias{US_TO_SECONDS}
\alias{JSTF_FORMAT_VERSION}
\alias{JSTF_DEFAULT_COMPRESSION}
\alias{STANDARD_SAMPLE_RATES}
\alias{AVQI_MIN_SV_DURATION_MS}
\alias{AVQI_MIN_CS_DURATION_MS}
\title{Package-level Constants for Signal Processing}
\usage{
PHONET_SAMPLE_RATE

DEFAULT_CHANNELS

SACC_SAMPLE_RATE

BROUHAHA_SAMPLE_RATE

MS_TO_SECONDS

US_TO_SECONDS

JSTF_FORMAT_VERSION

JSTF_DEFAULT_COMPRESSION

STANDARD_SAMPLE_RATES

AVQI_MIN_SV_DURATION_MS

AVQI_MIN_CS_DURATION_MS
}
\description{
This file defines package-level constants used throughout superassp for
signal processing, audio handling, and file format specifications.
}
\keyword{internal}
