% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trk_formants_tvwlp.R
\name{construct_output_path}
\alias{construct_output_path}
\title{Construct output file path}
\usage{
construct_output_path(input_file, extension, output_dir)
}
\description{
Construct output file path
}
\keyword{internal}
