% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_av_sptk_helpers.R
\name{create_aperiodicity_asspobj}
\alias{create_aperiodicity_asspobj}
\title{Convert SPTK C++ aperiodicity result to AsspDataObj}
\usage{
create_aperiodicity_asspobj(ap_result, windowShift)
}
\arguments{
\item{ap_result}{List returned from d4c_cpp}

\item{windowShift}{Frame shift in milliseconds}
}
\value{
AsspDataObj with aperiodicity track
}
\description{
Convert SPTK C++ aperiodicity result to AsspDataObj
}
\keyword{internal}
