% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptk_helpers.R
\name{create_formant_asspobj}
\alias{create_formant_asspobj}
\title{Convert Snack formant result to AsspDataObj}
\usage{
create_formant_asspobj(res, windowShift, numFormants)
}
\arguments{
\item{res}{List from snackf_cpp (fm, bw, sample_rate, n_frames)}

\item{windowShift}{Frame shift in milliseconds}

\item{numFormants}{Number of formants}
}
\value{
AsspDataObj with fm and bw tracks
}
\description{
Convert Snack formant result to AsspDataObj
}
\keyword{internal}
