% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_core.R
\name{create_json_track_obj}
\alias{create_json_track_obj}
\title{Create a JsonTrackObj}
\usage{
create_json_track_obj(
  results,
  function_name,
  file_path,
  sample_rate = NULL,
  audio_duration = NULL,
  beginTime = 0,
  endTime = 0,
  parameters = list()
)
}
\arguments{
\item{results}{List of results from lst_* function}

\item{function_name}{Name of the DSP function}

\item{file_path}{Original audio file path}

\item{sample_rate}{Audio sample rate in Hz}

\item{audio_duration}{Total audio duration in seconds}

\item{beginTime}{Start time in seconds}

\item{endTime}{End time in seconds}

\item{parameters}{Function parameters used}
}
\value{
Object of class JsonTrackObj
}
\description{
Create a JsonTrackObj
}
