% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptk_helpers.R
\name{create_snackp_asspobj}
\alias{create_snackp_asspobj}
\title{Convert Snack pitch result (4 tracks) to AsspDataObj}
\usage{
create_snackp_asspobj(res, windowShift)
}
\arguments{
\item{res}{List from snackp_cpp (f0, voicing, rms, acpeak, sample_rate, n_frames)}

\item{windowShift}{Frame shift in milliseconds}
}
\value{
AsspDataObj with f0, voicing, rms, acpeak tracks
}
\description{
Convert Snack pitch result (4 tracks) to AsspDataObj
}
\keyword{internal}
