% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptk_helpers.R
\name{create_spectrogram_asspobj}
\alias{create_spectrogram_asspobj}
\title{Convert CheapTrick C++ result to AsspDataObj}
\usage{
create_spectrogram_asspobj(ct_result, windowShift)
}
\arguments{
\item{ct_result}{List returned from cheap_trick_cpp}

\item{windowShift}{Frame shift in milliseconds}
}
\value{
AsspDataObj with 'sp' track (n_frames x fft_size/2+1 matrix)
}
\description{
Convert CheapTrick C++ result to AsspDataObj
}
\keyword{internal}
