% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat.R
\name{cs_wrap_praat_script}
\alias{cs_wrap_praat_script}
\title{A concurrency safe way to wrap a Praat script into an R function}
\usage{
cs_wrap_praat_script(
  praat_location,
  script_code_to_run,
  directory,
  return = c("last-argument", "info-window")
)
}
\arguments{
\item{praat_location}{path to the Praat executable}

\item{script_code_to_run}{Praat script to run}

\item{directory}{the full path of a directory set up by \link{make_dsp_environment}}

\item{return}{value to return. "last-argument" returns the last argument to the Praat script. "info-window" returns the contents of the Praat Info Window.}
}
\value{
see \code{return} argument
}
\description{
This function is an extension of the \href{https://github.com/tjmahr/tjm.praat/blob/bc45b8cb32694e5109986d008ef85dbdcf72bf44/R/main.R#L16]}{tjm.praat::wrap_praat_script} function
that wraps a Praat script into an R function, but with an extra demand to
specify a directory where the Praat script should be stored before execution.
This means that rather than having multiple  Praat script executing from the
same temporary directory (which is the case when using
\href{https://github.com/tjmahr/tjm.praat/blob/bc45b8cb32694e5109986d008ef85dbdcf72bf44/R/main.R#L16]}{tjm.praat::wrap_praat_script}, an environment is expected to have been
created using the \link{make_dsp_environment} already, and supplied to this
function via the \code{directory} argument. Any additional sound files or batches of sound files
will then only be available to the particular instance of Praat and DSP Praat functions can be called concurrently.
}
