% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff.R
\name{cut.AsspDataObj}
\alias{cut.AsspDataObj}
\title{Cut out a portion of an SSFF object}
\usage{
\method{cut}{AsspDataObj}(obj, where, n_preceeding, n_following)
}
\arguments{
\item{obj}{An in-memory object of class \code{AsspDataObj}.}

\item{where}{The time point (relative to the signal's total duration) where the cutout will be centered (0.0-1.0)}

\item{n_preceeding}{The (maximum) number of data samples to include in addition to, and \emph{before}, the sample closest to the \code{where} time point.}

\item{n_following}{The (maximum) number of data samples to include in addition to, and \emph{after}, the sample closest to the \code{where} time point.}
}
\value{
an \code{AsspDataObj} object that contains just the data samples at, and possibly surrounding, the \code{where} relative time reference in the input obj, but with time references relative to the timeline of the original object.
}
\description{
Cut out a portion of an SSFF object
}
