% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_deepformants.R
\name{deepformants_info}
\alias{deepformants_info}
\title{Get DeepFormants information}
\usage{
deepformants_info()
}
\value{
A list with information about:
\itemize{
\item Python path and version
\item PyTorch version
\item NumPy version
\item SciPy version
\item Pandas version
\item Numba version
\item DeepFormants module path
\item Available model files
}
}
\description{
Returns information about the DeepFormants installation and configuration.
}
\examples{
\dontrun{
if (deepformants_available()) {
  info <- deepformants_info()
  print(info)
}
}

}
\seealso{
\code{\link{install_deepformants}} to install DeepFormants dependencies
\code{\link{deepformants_available}} to check availability
}
