% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated-egg-functions}
\alias{deprecated-egg-functions}
\alias{trk_egg_f0_deprecated}
\title{Deprecated EGG Functions - Migrated to eggstract Package}
\usage{
trk_egg_f0_deprecated(...)
}
\description{
These functions have been migrated to the \strong{eggstract} package as part
of a package restructuring effort. The eggstract package provides a unified
API for electroglottographic (EGG) signal analysis with support for multiple
output formats (dataframe, SSFF, suggestion).

\strong{Migration Timeline:}
\itemize{
\item These wrapper functions will remain available until superassp v0.10.0
\item Expected removal: 6-12 months from 2025-10-30
\item Users should update their code to use eggstract functions directly
}
}
\keyword{internal}
