% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtrack_geoms.R
\name{.assp_time_col}
\alias{.assp_time_col}
\title{Locate the time column of a track table}
\usage{
.assp_time_col(data, time = "frame_time")
}
\arguments{
\item{data}{data.frame. Track table.}

\item{time}{Character. Preferred name of the time column.}
}
\value{
Character. Name of the column holding time in seconds.
}
\description{
Locate the time column of a track table
}
\keyword{internal}
