% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_pvoice_tremor.R
\name{.calculate_contour_hnr}
\alias{.calculate_contour_hnr}
\title{Calculate harmonicity (HNR) for contour sound}
\usage{
.calculate_contour_hnr(sound, min_tremor_freq, max_tremor_freq)
}
\description{
Calculate harmonicity (HNR) for contour sound
}
\keyword{internal}
