% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_avqi.R
\name{.calculate_cpps_pladdrr}
\alias{.calculate_cpps_pladdrr}
\title{Calculate CPPS using pladdrr}
\usage{
.calculate_cpps_pladdrr(sound)
}
\description{
Calculate CPPS using pladdrr
}
\keyword{internal}
