% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_python_pm_pavqi.R
\name{.calculate_ltas_slope_and_tilt_pladdrr}
\alias{.calculate_ltas_slope_and_tilt_pladdrr}
\title{Calculate LTAS slope and tilt using pladdrr}
\usage{
.calculate_ltas_slope_and_tilt_pladdrr(sound)
}
\description{
Calculate LTAS slope and tilt using pladdrr
}
\keyword{internal}
