% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_avqi.R
\name{.calculate_shimmer_pladdrr}
\alias{.calculate_shimmer_pladdrr}
\title{Calculate shimmer using pladdrr}
\usage{
.calculate_shimmer_pladdrr(sound)
}
\description{
Calculate shimmer using pladdrr
}
\keyword{internal}
