% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_voice_tremor.R
\name{.calculate_tremor_intensity}
\alias{.calculate_tremor_intensity}
\title{Calculate tremor intensity following Praat's tremIntIndex procedure}
\usage{
.calculate_tremor_intensity(
  contour_sound,
  tremor_pitch,
  slength,
  apply_correction = FALSE
)
}
\description{
Calculate tremor intensity following Praat's tremIntIndex procedure
}
\keyword{internal}
