% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_pvoice_tremor.R
\name{.extract_pitch_intensity_and_strength}
\alias{.extract_pitch_intensity_and_strength}
\title{Extract pitch intensity and strength from Pitch object}
\usage{
.extract_pitch_intensity_and_strength(pitch_object, max_freq, slength = NULL)
}
\description{
Extract pitch intensity and strength from Pitch object
}
\keyword{internal}
