% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_pavqi.R
\name{.extract_voiced_segments_pladdrr}
\alias{.extract_voiced_segments_pladdrr}
\title{Extract and concatenate voiced segments from continuous speech}
\usage{
.extract_voiced_segments_pladdrr(sound, version = "v3.01")
}
\description{
Extract and concatenate voiced segments from continuous speech
}
\keyword{internal}
