% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_pvoice_tremor.R
\name{.has_pitch_detrend}
\alias{.has_pitch_detrend}
\title{Check if pladdrr v4.0.14+ detrending is available}
\usage{
.has_pitch_detrend()
}
\description{
Check if pladdrr v4.0.14+ detrending is available
}
\keyword{internal}
