% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_pvoice_tremor.R
\name{.has_pp_values_from_sound}
\alias{.has_pp_values_from_sound}
\title{Check if pladdrr v4.0.14+ pp$get_values_from_sound is available}
\usage{
.has_pp_values_from_sound()
}
\description{
Check if pladdrr v4.0.14+ pp$get_values_from_sound is available
}
\keyword{internal}
