% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_pladdrr_psauce.R
\name{.iseli_alwan_correction}
\alias{.iseli_alwan_correction}
\title{Iseli & Alwan (2004) Formant Correction}
\usage{
.iseli_alwan_correction(f, fx, bx, fs)
}
\arguments{
\item{f}{Frequency of harmonic being measured (Hz)}

\item{fx}{Formant frequency (Hz)}

\item{bx}{Formant bandwidth (Hz)}

\item{fs}{Sampling frequency (Hz)}
}
\value{
Correction in dB to subtract from measured harmonic
}
\description{
Internal helper function implementing Iseli & Alwan formant correction.
Calculates how much a formant affects harmonic amplitude measurement.
}
\keyword{internal}
