% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_avqi.R
\name{.load_and_concatenate_sounds}
\alias{.load_and_concatenate_sounds}
\title{Load and concatenate sound objects}
\usage{
.load_and_concatenate_sounds(sounds)
}
\description{
Load and concatenate sound objects
}
\keyword{internal}
