% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnxruntime.R
\name{.ort_detect_platform}
\alias{.ort_detect_platform}
\title{Detect platform for ONNX Runtime download}
\usage{
.ort_detect_platform(gpu = FALSE)
}
\arguments{
\item{gpu}{Logical. Request GPU variant.}
}
\value{
List with os, arch, slug.
}
\description{
Detect platform for ONNX Runtime download
}
\keyword{internal}
