% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_onnxruntime.R
\name{.ort_resolve_lib_dir}
\alias{.ort_resolve_lib_dir}
\title{Resolve ONNX Runtime library directory}
\usage{
.ort_resolve_lib_dir()
}
\value{
Character path to library directory. May not exist.
}
\description{
Resolve ONNX Runtime library directory
}
\keyword{internal}
