% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_onnxruntime.R
\name{.ort_set_cached_path}
\alias{.ort_set_cached_path}
\title{Set cached path if install exists but option is unset}
\usage{
.ort_set_cached_path()
}
\description{
Set cached path if install exists but option is unset
}
\keyword{internal}
