% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_cpp_tvwlp.R
\name{.tvwlp_build_system}
\alias{.tvwlp_build_system}
\title{Build LP system matrix for L1 solver}
\usage{
.tvwlp_build_system(x, p, q, w = NULL)
}
\description{
Build LP system matrix for L1 solver
}
\keyword{internal}
