% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_cpp_tvwlp.R
\name{.tvwlp_filtfilt_iir}
\alias{.tvwlp_filtfilt_iir}
\title{Zero-phase IIR filter (forward-backward, direct-form II transposed)}
\usage{
.tvwlp_filtfilt_iir(b, a, x)
}
\description{
Zero-phase IIR filter (forward-backward, direct-form II transposed)
}
\keyword{internal}
