% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_cpp_tvwlp.R
\name{.tvwlp_qcp_wt}
\alias{.tvwlp_qcp_wt}
\title{Quasi-Closed Phase (QCP) weighting function}
\usage{
.tvwlp_qcp_wt(x, p, DQ = 0.7, PQ = 0.05, d = 1e-05, Nramp = 3L, gci_ins, fs)
}
\description{
Quasi-Closed Phase (QCP) weighting function
}
\keyword{internal}
