% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_cpp_tvwlp.R
\name{.tvwlp_resample_signal}
\alias{.tvwlp_resample_signal}
\title{Resample signal with length matching}
\usage{
.tvwlp_resample_signal(x, p, q)
}
\description{
Resample signal with length matching
}
\keyword{internal}
