% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_cpp_tvwlp.R
\name{.tvwlp_sedreams_gci_detection}
\alias{.tvwlp_sedreams_gci_detection}
\title{SEDREAMS GCI detection (requires fs = 8000)}
\usage{
.tvwlp_sedreams_gci_detection(wave, fs, f0_mean)
}
\description{
SEDREAMS GCI detection (requires fs = 8000)
}
\keyword{internal}
