% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_cpp_tvwlp.R
\name{.tvwlp_srh_estimate_pitch}
\alias{.tvwlp_srh_estimate_pitch}
\title{Frame-by-frame SRH pitch estimation}
\usage{
.tvwlp_srh_estimate_pitch(sig, fs, f0_min, f0_max)
}
\description{
Frame-by-frame SRH pitch estimation
}
\keyword{internal}
