% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_cpp_tvwlp.R
\name{.tvwlp_srh_pitch_tracking}
\alias{.tvwlp_srh_pitch_tracking}
\title{Two-pass SRH pitch tracking with range refinement}
\usage{
.tvwlp_srh_pitch_tracking(wave, fs, f0_min = 60L, f0_max = 600L)
}
\description{
Two-pass SRH pitch tracking with range refinement
}
\keyword{internal}
