% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_voice_tremor.R
\name{.undefined_atrem_results}
\alias{.undefined_atrem_results}
\title{Return undefined amplitude tremor results}
\usage{
.undefined_atrem_results(nanAsZero)
}
\description{
Return undefined amplitude tremor results
}
\keyword{internal}
