% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_voice_tremor.R
\name{.undefined_ftrem_results}
\alias{.undefined_ftrem_results}
\title{Return undefined frequency tremor results}
\usage{
.undefined_ftrem_results(nanAsZero)
}
\description{
Return undefined frequency tremor results
}
\keyword{internal}
