% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat_internal_lf.R
\name{.vat_lf_cont}
\alias{.vat_lf_cont}
\title{Generate one LF model pulse.}
\usage{
.vat_lf_cont(F0, fs, Ra, Rk, Rg, EE = 1)
}
\arguments{
\item{F0}{fundamental frequency (Hz)}

\item{fs}{sampling frequency (Hz)}

\item{Ra, Rk, Rg}{LF R-parameters}

\item{EE}{excitation strength}
}
\value{
numeric vector (one glottal-flow-derivative pulse)
}
\description{
Generate one LF model pulse.
}
