% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat_internal_lf.R
\name{.vat_rd2r}
\alias{.vat_rd2r}
\title{Convert Rd parameter to LF Ra/Rk/Rg.}
\usage{
.vat_rd2r(Rd, EE, F0)
}
\arguments{
\item{Rd}{shape descriptor}

\item{EE}{excitation strength}

\item{F0}{fundamental frequency (Hz)}
}
\value{
list with Ra, Rk, Rg
}
\description{
Convert Rd parameter to LF Ra/Rk/Rg.
}
