% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_r_voxit.R
\name{.voxit_alignment_metrics}
\alias{.voxit_alignment_metrics}
\title{Compute alignment-based metrics}
\usage{
.voxit_alignment_metrics(align_path, signal_metrics)
}
\description{
Compute alignment-based metrics
}
\keyword{internal}
