% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_r_voxit.R
\name{.voxit_build_jstf}
\alias{.voxit_build_jstf}
\title{Write JSTF output}
\usage{
.voxit_build_jstf(result, file_path, beginTime, endTime, minF, maxF)
}
\description{
Write JSTF output
}
\keyword{internal}
