% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_r_voxit.R
\name{.voxit_na_alignment_metrics}
\alias{.voxit_na_alignment_metrics}
\title{NA result for alignment metrics}
\usage{
.voxit_na_alignment_metrics()
}
\description{
NA result for alignment metrics
}
\keyword{internal}
