% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_r_voxit.R
\name{.voxit_pause_analysis}
\alias{.voxit_pause_analysis}
\title{Pause analysis helper}
\usage{
.voxit_pause_analysis(starts, ends, gentle_length)
}
\description{
Pause analysis helper
}
\keyword{internal}
