% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_r_voxit.R
\name{.voxit_rhythmic_complexity}
\alias{.voxit_rhythmic_complexity}
\title{Rhythmic complexity via LZ}
\usage{
.voxit_rhythmic_complexity(starts, ends, gentle_length)
}
\description{
Rhythmic complexity via LZ
}
\keyword{internal}
