% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_r_voxit.R
\name{.voxit_world_features}
\alias{.voxit_world_features}
\title{Compute WORLD-based signal metrics}
\usage{
.voxit_world_features(file_path, beginTime, endTime, minF, maxF)
}
\description{
Compute WORLD-based signal metrics
}
\keyword{internal}
