% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_r_voxit.R
\name{.voxit_write_jstf}
\alias{.voxit_write_jstf}
\title{Write JSTF output}
\usage{
.voxit_write_jstf(
  result,
  file_path,
  beginTime,
  endTime,
  minF,
  maxF,
  explicitExt,
  outputDirectory
)
}
\description{
Write JSTF output
}
\keyword{internal}
