% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_avaudio.R
\name{.write_wav_file}
\alias{.write_wav_file}
\title{Write WAV File}
\usage{
.write_wav_file(filename, samples, sample_rate, channels)
}
\arguments{
\item{filename}{Character; output file path}

\item{samples}{Integer vector; audio samples (s32le format)}

\item{sample_rate}{Integer; sample rate in Hz}

\item{channels}{Integer; number of channels}
}
\description{
Internal function to write audio samples to WAV file format.
}
\keyword{internal}
