% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{egg_migration_info_superassp}
\alias{egg_migration_info_superassp}
\title{Helper Function: EGG Migration Information}
\usage{
egg_migration_info_superassp()
}
\description{
Displays information about the migration of EGG functions from superassp to eggstract.
}
