% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat_internal_creak.R
\name{estimate_f0_simple}
\alias{estimate_f0_simple}
\title{Simple autocorrelation F0 estimate (per-sample via frame interpolation)}
\usage{
estimate_f0_simple(x, fs, F0min, F0max)
}
\description{
Simple autocorrelation F0 estimate (per-sample via frame interpolation)
}
\keyword{internal}
