% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_pharyngeal.R
\name{extract_ltas_peaks}
\alias{extract_ltas_peaks}
\title{Internal: Extract LTAS Peaks}
\usage{
extract_ltas_peaks(ltas, fmins, fmaxs)
}
\description{
Internal: Extract LTAS Peaks
}
\keyword{internal}
