% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_core.R
\name{extract_values_from_results}
\alias{extract_values_from_results}
\title{Extract values from results matching field schema}
\usage{
extract_values_from_results(results, field_names)
}
\arguments{
\item{results}{Results object}

\item{field_names}{Names of fields to extract}
}
\value{
List of values
}
\description{
Extract values from results matching field schema
}
\keyword{internal}
