% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat_internal_voice_quality.R
\name{find_amid_t}
\alias{find_amid_t}
\title{Find open-phase start and end times for QOQ}
\usage{
find_amid_t(glot_adj, Amid, Tz)
}
\description{
Find open-phase start and end times for QOQ
}
\keyword{internal}
