% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_c_assp_ksvfo.R
\name{fo}
\alias{fo}
\alias{foana}
\alias{fo_ksv}
\alias{ksvfo}
\title{Finds the f0 using the K.Schaefer-Vincent periodicity detection algorithm (Rcpp-optimized)}
\usage{
fo(
  listOfFiles = NULL,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  gender = "u",
  maxF = 600,
  minF = 50,
  minAmp = 50,
  maxZCR = 3000,
  toFile = TRUE,
  explicitExt = "fo",
  outputDirectory = NULL,
  assertLossless = NULL,
  logToFile = FALSE,
  convertOverwrites = FALSE,
  keepConverted = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{gender}{= \if{html}{\out{<code>}}  set gender-specific \ifelse{html}{\out{f<sub>o</sub>}}{\eqn{f_o}} ranges; \if{html}{\out{<code>}} may be:
"f\link{emale}" (80.0 - 640.0 Hz)
"m\link{ale}" (50.0 - 400.0 Hz)
"u\link{nknown}" (default; 50.0 - 600.0 Hz)}

\item{maxF}{= \if{html}{\out{<freq>}}: set maximum \ifelse{html}{\out{f<sub>o</sub>}}{\eqn{f_o}} value to \if{html}{\out{<freq>}} Hz (default: 500.0)}

\item{minF}{= \if{html}{\out{<freq>}}: set minimum \ifelse{html}{\out{f<sub>o</sub>}}{\eqn{f_o}} value to \if{html}{\out{<freq>}} Hz (default: 50.0)}

\item{minAmp}{= \if{html}{\out{<amp>}}: set amplitude threshold for voiced samples to \if{html}{\out{<amp>}} (default: 100)}

\item{maxZCR}{maximum zero crossing rate in Hz (for voicing detection)}
}
\value{
The number of successfully written files (if \code{toFile=TRUE}), or a vector of \code{AsspDataObj} objects (if \code{toFile=FALSE}).
}
\description{
Applies Schäefer-Vincent periodicity analysis \insertCite{Schäfer-Vincent.1983.10.1159/000261691}{superassp}
to find \ifelse{html}{\out{f<sub>o</sub>}}{\eqn{f_o}} (in Hz)
along signals listed in \code{listOfFiles}. Input signals not in a file format natively
supported will be converted before the autocorrelation functions are
computed. The conversion process will display warnings about input files
that are not in known losslessly encoded formats.
}
\details{
The results will be will be written to an SSFF formated file with the base
name of the input file and extension \emph{.fo} in a track \emph{fo\link{Hz}}.

The function is a re-write of the \link[wrassp:ksvF0]{wrassp::ksvF0} function, but
with media pre-conversion, better checking of preconditions such as the
input file existence, structured logging, and the use of a more modern
framework for user feedback. This version includes Rcpp optimizations
for improved performance on large batches of files.

Optionally, location and type of the signal extrema on
which the \ifelse{html}{\out{f<sub>o</sub>}}{\eqn{f_o}} data are based, may be stored in a label
file. The name of this file will consist of the basename of the \code{.fo} file and the extension '.prd'.

The native file type of this function is "wav" files (in "pcm_s16le"
format), SUNs "au", NIST, or CSL formats (kay or NSP extension). Input
signal conversion, when needed, is done by
\href{https://ffmpeg.org/libavcodec.html}{libavcodec} and the excellent \link[av:encoding]{av::av_audio_convert}
wrapper function
}
\note{
This function is not considered computationally expensive enough to require caching of
results if applied to many signals. However, if the number of signals it will be applied to
is \emph{very} long, then caching of results may be warranted.
}
\examples{
# get path to audio file
path2wav <- list.files(system.file("samples","sustained", package = "superassp"), pattern = glob2rx("a1.wav"), full.names = TRUE)

# calculate fundamental frequency contour
res <- trk_ksvfo(path2wav, toFile=FALSE)

# plot the fundamental frequency contour
plot(seq(0,numRecs.AsspDataObj(res) - 1) / rate.AsspDataObj(res) +
      attr(res, 'startTime'),
    res[["fo[Hz]"]],
    type='l',
    xlab='time (s)',
    ylab=expression(paste(f[o]," frequency (Hz)")))

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{pitch}} for a tracker of pitch
}
\author{
Raphael Winkelmann

Lasse Bombien

Fredrik Nylén
}
