% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat_internal_se_vq.R
\name{gci_creak_postproc}
\alias{gci_creak_postproc}
\title{Remove false GCIs in creaky regions}
\usage{
gci_creak_postproc(GCI, creak, search_reg, rep, remove_thresh, rep_num)
}
\description{
Remove false GCIs in creaky regions
}
\keyword{internal}
