% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disvoice_utils.R
\name{get_audio_duration}
\alias{get_audio_duration}
\title{Get Audio Duration}
\usage{
get_audio_duration(audio_path)
}
\arguments{
\item{audio_path}{Path to audio file}
}
\value{
Duration in seconds
}
\description{
Get duration of audio file in seconds using av.
}
\keyword{internal}
