% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdata_extensions.R
\name{get_extension}
\alias{get_extension}
\title{Get the (default) extension for an SSFF producing function or a signal file}
\usage{
get_extension(x)
}
\arguments{
\item{x}{The name of function defined to output trackdata, or the full path to a file that can be read using \link[wrassp:read.AsspDataObj]{wrassp::read.AsspDataObj}.}
}
\value{
A string indicating the default file extension of the SSFF generating function, or the file extension of the signal file.
}
\description{
Get the (default) extension for an SSFF producing function or a signal file
}
\examples{
superassp:::get_extension("trk_formant_forest")
superassp:::get_extension("trk_formant_burg")

}
\keyword{internal}
