% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstf_io.R
\name{get_jstf_extension}
\alias{get_jstf_extension}
\title{Get JSTF extension for function}
\usage{
get_jstf_extension(function_name)
}
\arguments{
\item{function_name}{Name of lst_* function}
}
\value{
File extension (without dot)
}
\description{
Get JSTF extension for function
}
\examples{
superassp:::get_jstf_extension("lst_avqi")  # "avq"
superassp:::get_jstf_extension("lst_dsi")  # "dsi"

}
\keyword{internal}
