% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_io.R
\name{get_jstf_extensions}
\alias{get_jstf_extensions}
\title{Get list of JSTF extensions}
\usage{
get_jstf_extensions()
}
\value{
Character vector of JSTF file extensions
}
\description{
Get list of JSTF extensions
}
\keyword{internal}
