% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat_internal_se_vq.R
\name{get_mbs_gci_intervals}
\alias{get_mbs_gci_intervals}
\title{GCI search intervals from MBS}
\usage{
get_mbs_gci_intervals(MBS, fs, T0mean, F0max)
}
\description{
GCI search intervals from MBS
}
\keyword{internal}
