% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdata_extensions.R
\name{get_outputType}
\alias{get_outputType}
\title{Get the return format of a wrassp/superassp speech signal processing function}
\usage{
get_outputType(x, package = "superassp")
}
\arguments{
\item{x}{The name of a speech signal processing function that are defined in the superassp or wrassp packages.}

\item{package}{The name of the package where the function is defined.}
}
\value{
Either "SSFF" or "list".
}
\description{
Speech signal processing functions in superassp may return an SSFF track with one or more columns, or a simple, not nested list. This function may be used to learn the output type of a specific function.
}
\examples{
superassp:::get_outputType("trk_formant_forest")
superassp:::get_outputType("lst_avqi")
superassp:::get_outputType("trk_formant_burg")

}
\keyword{internal}
