% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pladdrr_helpers.R
\name{get_pladdrr_ptr}
\alias{get_pladdrr_ptr}
\title{Get external pointer from pladdrr R6 object}
\usage{
get_pladdrr_ptr(obj)
}
\arguments{
\item{obj}{pladdrr R6 object (Sound, Pitch, Formant, etc.)}
}
\value{
External pointer to the Praat C object
}
\description{
Extract the underlying C pointer from a pladdrr R6 object.
Used internally for passing objects to direct API functions.
}
\details{
pladdrr uses R6 function factory pattern where all objects have a .xptr field
containing the C pointer.
}
\examples{
\dontrun{
sound <- pladdrr::Sound("audio.wav")
ptr <- get_pladdrr_ptr(sound)
}

}
\keyword{internal}
