% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff.R
\name{harmonics}
\alias{harmonics}
\title{Compute the harmonic frequency structure from f0 measurements}
\usage{
harmonics(track, column = "f0", n = 5, explicitExt = "har", toFile = TRUE)
}
\arguments{
\item{track}{An f0 track, either as an SSFF object or as the name of an SSFF formatted file. It is recommended that the}

\item{column}{The name of the column to use as an f0 track.}

\item{n}{The number of harmonics to compute.}

\item{explicitExt}{The output file extension.}

\item{toFile}{boolean;Should the SSFF track be returned or stored on disc?}
}
\value{
The SSFF track object, if required. \code{NULL} otherwise.
}
\description{
This function takes a pre-computed f0 track and derive \code{n} harmonic tracks
from it so that the vector of f0 values are now a matrix with \code{n} columns.
Each column then encode the \code{n}th harmonic values.
}
\details{
The stored harmonic frequencies are simply multiples of the fundamental frequency (f0) track, and not derived independently from the speech signal. Therefore, errors in the frequency tracking of the f0 signal will be carried over to these tracks.
The primary use case for the track is to have have estimates of the harmonic frequencies to visualize harmonic frequency (n*f~0~) against harmonic amplitude (L~1-n~) .
}
