% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_ssff.R
\name{harvest}
\alias{harvest}
\title{Compute f0 using the Harvest algorithm}
\usage{
harvest(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 70,
  maxF = 200,
  explicitExt = "hf0",
  outputDirectory = NULL,
  toFile = TRUE
)
}
\arguments{
\item{listOfFiles}{Vector of file paths to process}

\item{beginTime}{Start time in seconds (default: 0.0)}

\item{endTime}{End time in seconds (default: 0.0 = end of file)}

\item{windowShift}{Frame shift in milliseconds (default: 10.0)}

\item{minF}{Minimum F0 in Hz (default: 60.0)}

\item{maxF}{Maximum F0 in Hz (default: 400.0)}

\item{explicitExt}{Output file extension (default: "f0")}

\item{outputDirectory}{Output directory (default: NULL = same as input)}

\item{toFile}{Write results to file (default: TRUE)}
}
\value{
An SSFF track object containing two tracks (f0 and corr) that are
either returned (toFile == FALSE) or stored on disk.
}
\description{
The Harvest algorithm \insertCite{Morise.2017.10.21437/interspeech.2017-68}{superassp} was developed for the WORLD vocoder \insertCite{MORISE.2016.10.1587/transinf.2015edp7457}{superassp} and aims to obtain a reliable F0 contour and reduce erroneously identified voice frames.
}
\details{
The algorithm consists of two steps. In the first step, the algorithm uses fundamental component extraction by many band-pass filters with different center frequencies and obtains the basic f0 candidates from filtered signals. In the second step, basic f0 candidates are refined and scored by using the instantaneous frequency, and then several f0 candidates in each frame are estimated.
}
\references{
\insertAllCited{}
}
