% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat_ssff.R
\name{have_praat}
\alias{have_praat}
\title{A simple check of a presence of a Praat executable}
\usage{
have_praat(praat_path = NULL)
}
\arguments{
\item{praat_path}{A character string containing the path to the executable that the function was able to find (or the executable that the function was able to verify the existance of), or NULL if no Praat executable was found and verified.}
}
\value{
A boolean indicating whether the Praat executable could be found or not.
}
\description{
A simple check of a presence of a Praat executable
}
