% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_core.R
\name{infer_field_schema}
\alias{infer_field_schema}
\title{Infer field schema from results}
\usage{
infer_field_schema(results)
}
\arguments{
\item{results}{List of results}
}
\value{
Named character vector of field types
}
\description{
Infer field schema from results
}
\keyword{internal}
