% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxit_dsp_exports.R
\name{interp1_linear_cpp}
\alias{interp1_linear_cpp}
\title{Linear interpolation}
\usage{
interp1_linear_cpp(xp, fp, xi)
}
\arguments{
\item{xp}{X positions of known points}

\item{fp}{Y values at known points (same length as xp)}

\item{xi}{X positions to interpolate at}
}
\value{
Interpolated Y values at xi
}
\description{
Linear interpolation at arbitrary points.
}
