% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dysprosody_intsint.R
\name{intsint}
\alias{intsint}
\title{Run INTSINT algorithm on MOMEL targets}
\usage{
intsint(targets)
}
\arguments{
\item{targets}{data.frame with time (frames) and frequency (Hz) from momel()}
}
\value{
list(targets_df, range, key) where targets_df has columns
time_sec, tone, target_hz, estimate_hz
}
\description{
Run INTSINT algorithm on MOMEL targets
}
