% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_avaudio.R
\name{is_avaudio}
\alias{is_avaudio}
\title{Check if Object is AVAudio}
\usage{
is_avaudio(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical; TRUE if x is an AVAudio object
}
\description{
Check if Object is AVAudio
}
