% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_integration_example.R
\name{json_track_integration}
\alias{json_track_integration}
\title{Example: Integrating JSTF with lst_* Functions}
\description{
This file demonstrates how to add toFile support to lst_* functions
using the JSON Track Format (JSTF).
}
\keyword{internal}
