% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superassp_rfcana.R
\name{larana}
\alias{larana}
\title{Linear Prediction analysis using autocorrelation the Durbin recursion and log area ratio output (Rcpp-optimized)}
\usage{
larana(listOfFiles = NULL,
  beginTime = 0.0,
  centerTime = FALSE,
  endTime = 0.0,
  windowShift = 5.0,
  windowSize = 20.0,
  effectiveLength = TRUE,
  window = 'BLACKMAN',
  order = 0,
  preemphasis = -0.95,
  toFile = TRUE,
  explicitExt=NULL,
  outputDirectory = NULL,
  assertLossless = NULL,
  logToFile = FALSE,
  keepConverted=FALSE,
  verbose = TRUE)
}
\value{
The number of successfully written files (if \code{toFile=TRUE}), or a vector of \code{AsspDataObj} objects (if \code{toFile=FALSE}).
}
\description{
Linear Prediction analysis of \code{listOfFiles} using the
autocorrelation method and the Durbin recursion.
This program calculates the RMS amplitudes of the input
and residual signal in dB and, and log area ratios using algorithms implmented in \emph{libassp} \insertCite{s5h}{superassp}.
Input signals not in a file format natively
supported will be converted before the autocorrelation functions are
computed. The conversion process will display warnings about input files
that are not in known losslessly encoded formats.
}
\details{
The results will be will be written to an SSFF formated file with the base
name of the input file and extension \code{.lar} with tracks \code{RMS[dB]},\code{gain[dB]},and \code{LAR}.

The function is a re-write of the \link[wrassp:rfcana]{wrassp::rfcana} function with log area ratio output, but
with media pre-conversion, better checking of preconditions such as the
input file existence, structured logging, and the use of a more modern
framework for user feedback. This version includes Rcpp optimizations
for improved performance on large batches of files.

The native file type of this function is "wav" files (in "pcm_s16le"
format), SUNs "au", NIST, or CSL formats (kay or NSP extension). Input
signal conversion, when needed, is done by
\href{https://ffmpeg.org/libavcodec.html}{libavcodec} and the excellent \link[av:encoding]{av::av_audio_convert}
wrapper function
}
\note{
This function is not considered computationally expensive enough to require caching of
results if applied to many signals. However, if the number of signals it will be applied to
is \emph{very} large, then caching of results may be warranted.
}
\examples{
# get path to audio file
path2wav <- list.files(system.file("samples","sustained", package = "superassp"), pattern = glob2rx("a1.wav"), full.names = TRUE)

# perform linear prediction analysis
res <- rfcana(path2wav, toFile=FALSE)

# plot reflection coefficients
matplot(seq(0,numRecs.AsspDataObj(res) - 1) / rate.AsspDataObj(res) +
          attr(res, 'startTime'),
        res$LAR,
        type='l',
        xlab='time (s)',
        ylab='Log area ratios')

}
\references{
\insertAllCited{}
}
\seealso{
\link[wrassp:rfcana]{wrassp::rfcana}

\link{AsspWindowTypes}

\link[av:encoding]{av::av_audio_convert}
}
\author{
Raphael Winkelmann

Lasse Bombien

Fredrik Nylén
}
