% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtorch_deepformants.R
\name{load_deepformants_estimator}
\alias{load_deepformants_estimator}
\title{Load bundled DeepFormants Estimator (MLP) weights}
\usage{
load_deepformants_estimator()
}
\description{
Load bundled DeepFormants Estimator (MLP) weights
}
\keyword{internal}
