% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtorch_deepformants.R
\name{load_deepformants_tracker}
\alias{load_deepformants_tracker}
\title{Load bundled DeepFormants Tracker (LSTM) weights}
\usage{
load_deepformants_tracker()
}
\description{
Load bundled DeepFormants Tracker (LSTM) weights
}
\keyword{internal}
