% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_python_opensmile_ComParE_2016.R
\name{lst_ComParE_2016_python}
\alias{lst_ComParE_2016_python}
\title{Compute the ComParE 2016 openSMILE feature set (Python Implementation)}
\usage{
lst_ComParE_2016_python(listOfFiles, ...)
}
\description{
Python implementation - used as fallback
}
\keyword{internal}
