% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cpp_opensmile_generic.R
\name{lst_eGeMAPS_cpp}
\alias{lst_eGeMAPS_cpp}
\title{Compute eGeMAPS Features (C++ Implementation)}
\usage{
lst_eGeMAPS_cpp(file, beginTime = 0, endTime = 0, verbose = FALSE)
}
\arguments{
\item{file}{Path to audio file}

\item{beginTime}{Start time in seconds (default: 0)}

\item{endTime}{End time in seconds (default: 0 = end of file)}

\item{verbose}{Print processing information (default: FALSE)}
}
\value{
Named list with 88 eGeMAPS features
}
\description{
Extracts Extended Geneva Minimalistic Acoustic Parameter Set (eGeMAPS) features
using OpenSMILE C++ library.
}
\keyword{internal}
