% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_python_opensmile_eGeMAPS.R
\name{lst_eGeMAPS_python}
\alias{lst_eGeMAPS_python}
\title{Compute the eGeMAPS openSMILE feature set (Python Implementation)}
\usage{
lst_eGeMAPS_python(listOfFiles, ...)
}
\description{
Python implementation - used as fallback
}
\keyword{internal}
