% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_python_opensmile_emobase.R
\name{lst_emobase_python}
\alias{lst_emobase_python}
\title{Compute the emobase openSMILE feature set (Python Implementation)}
\usage{
lst_emobase_python(listOfFiles, ...)
}
\description{
Python implementation - used as fallback
}
\keyword{internal}
