% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat.R
\name{make_dsp_environment}
\alias{make_dsp_environment}
\title{A utility function to make a concurrency-safe DSP environment}
\usage{
make_dsp_environment()
}
\value{
The full path of the constructed directory (string).
}
\description{
The function will make a unique directory associated with a
particular speech signal file, which will then be unique ans safe for calling
Praat DSP functions concurrently.
}
\details{
Teardown of the environment should be done using the \link{clear_dsp_environment} function,
which should then be called using the same signal file name. The file path is used "as is" when
creating the concurrency safe environment, so the user should make sure to do any file path
normalization prior to using this function and \link{clear_dsp_environment} so that the unique
identifier of the environment being constructed and tore down are ensured to be the same.
}
\seealso{
clear_dsp_environment
}
