% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superassp_fileHelper.R, R/superassp_rmsana.R
\name{makeOutputDirectory}
\alias{makeOutputDirectory}
\title{Create and validate output directory}
\usage{
makeOutputDirectory(outputDirectory, logToFile = FALSE, funName)

makeOutputDirectory(outputDirectory, logToFile = FALSE, funName)
}
\arguments{
\item{outputDirectory}{Character path to output directory}

\item{logToFile}{Logical, whether to create log file}

\item{funName}{Character, name of calling function}
}
\description{
Create and validate output directory

Create and validate output directory
}
\keyword{internal}
