% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_methods.R
\name{merge_json_tracks}
\alias{merge_json_tracks}
\title{Merge multiple JsonTrackObj files}
\usage{
merge_json_tracks(...)
}
\arguments{
\item{...}{JsonTrackObj objects to merge}
}
\value{
Merged JsonTrackObj
}
\description{
Combines multiple JsonTrackObj objects from the same audio file.
Useful when processing different time slices separately.
}
