% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrassp_mhsF0.R
\name{mhsfo}
\alias{mhsfo}
\alias{mhspitch}
\alias{fo_mhs}
\title{Pitch analysis using a Modified Harmonic Sieve}
\usage{
mhsfo(
  listOfFiles = NULL,
  beginTime = 0,
  centerTime = FALSE,
  endTime = 0,
  windowShift = 5,
  gender = "u",
  maxF = 600,
  minF = 50,
  minAmp = 50,
  minAC1 = 0.25,
  minRMS = 18,
  maxZCR = 3000,
  minProb = 0.52,
  plainSpectrum = FALSE,
  toFile = TRUE,
  explicitExt = "pit",
  outputDirectory = NULL,
  assertLossless = NULL,
  logToFile = FALSE,
  convertOverwrites = FALSE,
  keepConverted = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{listOfFiles}{vector of file paths to be processed by function}

\item{beginTime}{the time point (in seconds) of the start of the analysed
interval. A NULL or 0 is interpreted as the start of the signal file.
If a vector of time points is supplied, the length of that vector needs
to correspond with the length of \code{listOfFiles}.}

\item{endTime}{the time point (in seconds) of the end of the analysed
interval. A NULL or 0 is interpreted as the end of the signal file.
If a vector of time points is supplied, the length of that vector needs
to correspond with the length of \code{listOfFiles}.}

\item{windowShift}{the amount of time (in ms) that the analysis window will
be shifted between analysis frames}

\item{gender}{= \if{html}{\out{<code>}}  set gender-specific \ifelse{html}{\out{f<sub>o</sub>}}{\eqn{f_o}} ranges; \if{html}{\out{<code>}} may be:
"f\link{emale}" (80.0 - 640.0 Hz)
"m\link{ale}" (50.0 - 400.0 Hz)
"u\link{nknown}" (default; 50.0 - 600.0 Hz)}

\item{maxF}{= \if{html}{\out{<freq>}}: set maximum \ifelse{html}{\out{f<sub>o</sub>}}{\eqn{f_o}} value to \if{html}{\out{<freq>}} Hz (default: 500.0)}

\item{minF}{= \if{html}{\out{<freq>}}: set minimum \ifelse{html}{\out{f<sub>o</sub>}}{\eqn{f_o}} value to \if{html}{\out{<freq>}} Hz (default: 50.0)}

\item{minAmp}{= \if{html}{\out{<amp>}}:  minimum signal amplitude (default: 50)}

\item{minAC1}{= \if{html}{\out{<freq>}}: minimum 1st correlation coefficient (default: 0.250)}

\item{minRMS}{= \if{html}{\out{<num>}}:  minimum RMS amplitude in dB (default: 18.0)}

\item{maxZCR}{= \if{html}{\out{<freq>}}: maximum zero crossing rate in Hz (default: 3000)}

\item{minProb}{= \if{html}{\out{<num>}}: minimum quality value of \ifelse{html}{\out{f<sub>o</sub>}}{\eqn{f_o}} fit (default: 0.520)}

\item{toFile}{Should the function write the results to a file, with the
(default) file extension \emph{.acf} (\code{TRUE}) or returned as a list of
\link{AsspDataObj} objects (\code{FALSE})?}

\item{explicitExt}{by default an \emph{.acf} file extension will be used when
result files are written (\code{toFile=TRUE}), but the file extension can be
set to something else using this function argument.}

\item{outputDirectory}{directory in which output files are stored. Defaults
to NULL which means that the result file will be stored in the same
directory as the input file.}

\item{assertLossless}{an optional list of file extensions that the user wants to assert
contains losslessly encoded signals data.}

\item{logToFile}{whether to log commands to a separate logfile in the
\code{outputDirectory}. Logging will otherwise be in the \code{acfana} logging
namespace of \link{logger} and will be put wherever this namespace is defined to place its output.
See \link[logger:log_appender]{logger::log_appender} for details.}

\item{verbose}{display verbose information about processing steps taken, as
well as progress bars.}
}
\value{
The number of successfully written files (if \code{toFile=TRUE}), or a vector of \code{AsspDataObj} objects (if \code{toFile=FALSE}).
}
\description{
This function finds the pitch (in Hz) along files in \if{html}{\out{<listOfFile>}} using
Michel Scheffers' Modified Harmonic Sieve algorithm implmented in \emph{libassp} \insertCite{s5h}{superassp}.
Input signals not in a file format natively
supported will be converted before the autocorrelation functions are
computed. The conversion process will display warnings about input files
that are not in known losslessly encoded formats.
}
\details{
The results will be will be written to an SSFF formated file with the base
name of the input file and extension \emph{.pit} in a track \emph{pitch}.

The function is a re-write of the \link[wrassp:mhsF0]{wrassp::mhsF0} function, but
with media pre-conversion, better checking of preconditions such as the
input file existance, structured logging, and the use of a more modern
framework for user feedback.

The native file type of this function is "wav" files (in "pcm_s16le"
format), SUNs "au", NIST, or CSL formats (kay or NSP extension). Input
signal conversion, when needed, is done by
\href{https://ffmpeg.org/libavcodec.html}{libavcodec} and the excellent \link[av:encoding]{av::av_audio_convert}
wrapper function
}
\note{
This function is not considered computationally expensive enough to require caching of
results if applied to many signals. However, if the number of signals it will be applied to
is \emph{very} long, then caching of results may be warranted.
}
\examples{
# get path to audio file
path2wav <- list.files(system.file("samples","sustained", package = "superassp"), pattern = glob2rx("a1.wav"), full.names = TRUE)

# calculate short-term autocorrelation
res <- fo_mhs(path2wav, toFile=FALSE)

# plot fundamental frequency contour
plot(seq(0,numRecs.AsspDataObj(res) - 1) / rate.AsspDataObj(res) +
       attr(res, 'startTime'),
     res[["pitch[Hz]"]],
     type='l',
     xlab='time (s)',
     ylab="Pitch (Hz)")

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ksv_fo}} for an alternative algorithm for tracking the fundamental frequency \ifelse{html}{\out{f<sub>o</sub>}}{\eqn{f_o}}.
}
\author{
Raphael Winkelmann

Lasse Bombien

Fredrik Nylén
}
