% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dysprosody_momel.R
\name{momel}
\alias{momel}
\title{Run MOMEL algorithm on F0 values}
\usage{
momel(
  f0_values,
  window_length = 30L,
  min_f0 = 60,
  max_f0 = 750,
  max_error = 1.04,
  reduced_window_length = 20L,
  minimal_distance = 5,
  minimal_frequency_ratio = 0.05
)
}
\arguments{
\item{f0_values}{numeric vector of F0 in Hz (10ms frames), 0 = unvoiced}

\item{window_length}{window in samples for cible (default 30ms / 10ms = 3)}

\item{min_f0}{minimum F0 in Hz}

\item{max_f0}{maximum F0 in Hz}

\item{max_error}{maximum error ratio}

\item{reduced_window_length}{window for reduction}

\item{minimal_distance}{minimal distance in frames}

\item{minimal_frequency_ratio}{minimal frequency ratio for merging}
}
\value{
data.frame with columns time (frames, 0-based) and frequency (Hz)
}
\description{
Run MOMEL algorithm on F0 values
}
