% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_ssff.R
\name{nonopt_swipe}
\alias{nonopt_swipe}
\title{Compute f0 using the SWIPE algorithm (Non-optimized Python version)}
\usage{
nonopt_swipe(
  listOfFiles = NULL,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 70,
  maxF = 200,
  voicing.threshold = 0.3,
  explicitExt = "swi",
  outputDirectory = NULL,
  toFile = TRUE,
  assertLossless = NULL,
  logToFile = FALSE,
  keepConverted = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{listOfFiles}{A vector of file paths to wav files.}

\item{beginTime}{The start time of the section of the sound file that should be processed.}

\item{endTime}{The end time of the section of the sound file that should be processed.}

\item{windowShift}{The measurement interval (frame duration), in seconds.}

\item{minF}{Candidate f0 frequencies below this frequency will not be considered.}

\item{maxF}{Candidates above this frequency will be ignored.}

\item{voicing.threshold}{Voice/unvoiced threshold. Default is 0.3.}

\item{explicitExt}{File extension for output files}

\item{outputDirectory}{Output directory path (NULL for same as input)}

\item{toFile}{Write output to file (TRUE) or return object (FALSE)}

\item{conda.env}{The name of the conda environment in which Python and its
required packages are stored. Please make sure that you know what you are
doing if you change this. Defaults to \code{NULL}, which means that the default enviroment or the environment set in the
\code{RETICULATE_PYTHON} environment variable will be used.}
}
\value{
An SSFF track object containing two tracks (f0 and pitch) that are either returned (toFile == FALSE) or stored on disk.
}
\description{
\code{\link[lifecycle]{lifecycle-superseded}}

This function has been superseded by \code{\link{swipe_cpp}}, which provides a native
C++ implementation that is 2-3x faster and does not require Python dependencies.
This Python-based version is retained for compatibility but is no longer exported.

This function takes a sound file and computes f$_0$ and an estimate of pitch
using the Sawtooth Waveform Inspired Pitch Estimator (SWIPE) algorithm \insertCite{Camacho.2008.10.1121/1.2951592}{superassp}.
}
\details{
The implementation of SWIPE in the Speech Signal Processing Toolkit (SPTK) \insertCite{sptkspeech}{superassp} is used, and called via its Python interface and the \link{retiulate} R package to compute the signal track.
Therefore, the user will have to make sure that a python environment is present and can be attached by the \link[reticulate:reticulate]{reticulate::reticulate}. An anaconda environment is recommended, and can set up by the user by a setup procedure that involve at least these commands:

\if{html}{\out{<div class="sourceCode">}}\preformatted{conda create conda create --prefix -n pysuperassp python=3.8 
conda activate pysuperassp
pip install librosa
pip install pysptk
}\if{html}{\out{</div>}}

to make the functionality that this function requires available.
}
\references{
\insertAllCited{}
}
\keyword{internal}
