% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pladdrr_pharyngeal.R
\name{normalize_bandwidth}
\alias{normalize_bandwidth}
\title{Internal: Normalize Bandwidth}
\usage{
normalize_bandwidth(bw, f, UNDEFINED)
}
\description{
Internal: Normalize Bandwidth
}
\keyword{internal}
