% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trk_formants_tvwlp.R
\name{numpy_to_assp_dataobj}
\alias{numpy_to_assp_dataobj}
\title{Convert NumPy array to AsspDataObj}
\usage{
numpy_to_assp_dataobj(
  data,
  track_names,
  track_format,
  sample_rate,
  frame_shift,
  track_description
)
}
\description{
Convert NumPy array to AsspDataObj
}
\keyword{internal}
