% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_onnxruntime.R
\name{onnxruntime_path}
\alias{onnxruntime_path}
\title{Get ONNX Runtime library path}
\usage{
onnxruntime_path()
}
\value{
Character. Path to the onnxruntime library directory.
}
\description{
Get ONNX Runtime library path
}
