% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cpp_opensmile_generic.R
\name{opensmile_extract_generic}
\alias{opensmile_extract_generic}
\title{Generic OpenSMILE Feature Extraction using C++}
\usage{
opensmile_extract_generic(
  file,
  config_name,
  config_dir,
  feature_set_name = "features",
  beginTime = 0,
  endTime = 0,
  verbose = FALSE
)
}
\arguments{
\item{file}{Path to audio file}

\item{config_name}{Name of the config file (without extension)}

\item{config_dir}{Directory path relative to inst/opensmile/config}

\item{feature_set_name}{Name for verbose output}

\item{beginTime}{Start time in seconds}

\item{endTime}{End time in seconds}

\item{verbose}{Print processing information}
}
\value{
Named list with acoustic features
}
\description{
Internal function that extracts OpenSMILE features using the C++ implementation
}
\keyword{internal}
