% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{opensmile_gemaps_cpp}
\alias{opensmile_gemaps_cpp}
\title{OpenSMILE GeMAPS Feature Extraction (C++ Implementation)}
\usage{
opensmile_gemaps_cpp(audio_obj, config_file, verbose = FALSE)
}
\arguments{
\item{audio_obj}{An AsspDataObj containing audio data}

\item{config_file}{Path to OpenSMILE configuration file}

\item{verbose}{Print processing information (default: FALSE)}
}
\value{
Named list with 62 GeMAPS features
}
\description{
Wrapper function for GeMAPS feature extraction
}
\keyword{internal}
