% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_phonet.R
\name{phonet_info}
\alias{phonet_info}
\title{Get Phonet information}
\usage{
phonet_info()
}
\value{
A list with information about:
\itemize{
\item Python path and version
\item TensorFlow version
\item Keras version
\item NumPy version
\item SciPy version
\item Numba version
\item Phonet version and location
\item Available phonological classes
}
}
\description{
Returns information about the Phonet installation and configuration.
}
\examples{
\dontrun{
if (phonet_available()) {
  info <- phonet_info()
  print(info)
}
}

}
\seealso{
\code{\link{install_phonet}} to install Phonet dependencies
\code{\link{phonet_available}} to check availability
}
