% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pladdrr_helpers.R
\name{pladdrr_df_to_superassp}
\alias{pladdrr_df_to_superassp}
\title{Convert pladdrr data frame to superassp format}
\usage{
pladdrr_df_to_superassp(df, type = "pitch")
}
\arguments{
\item{df}{Data frame from pladdrr (e.g., pitch$as_data_frame())}

\item{type}{Type of data: "pitch", "formant", "intensity", etc.}
}
\value{
Data frame in superassp format
}
\description{
Converts pladdrr's long-format data frames to the wide format expected
by superassp/AsspDataObj.
}
\details{
pladdrr returns data in long format:
\itemize{
\item Pitch: time, frequency
\item Formant: time, formant, frequency, bandwidth
\item Intensity: time, intensity
}

superassp expects wide format:
\itemize{
\item Pitch: time, pitch
\item Formant: time, F1, F2, F3, F4, F5, B1, B2, B3, B4, B5
\item Intensity: time, intensity
}
}
\keyword{internal}
