% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat_ssff.R
\name{praat_formantpath_burg}
\alias{praat_formantpath_burg}
\title{Formant estimation using the FormantPath functionality of Praat}
\usage{
praat_formantpath_burg(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  numFormants = 5,
  maxFormantHz = 5500,
  windowSize = 30,
  preemphasis = 50,
  ceilingStepSize = 0.05,
  stepsUpDown = 4,
  windowShape = "Gaussian1",
  relativeWidth = 1,
  spectWindowShape = "Gaussian",
  spectResolution = 40,
  toFile = TRUE,
  explicitExt = "pfm",
  outputDirectory = NULL,
  verbose = FALSE,
  praat_path = NULL
)
}
\arguments{
\item{listOfFiles}{a vector of wav file paths to be processed by function.}

\item{beginTime}{the time where processing should end (in s) The default is 0 (zero) which means that the computation of formants will start at the start of the sound file.}

\item{endTime}{the time where processing should end (in s) The default is 0 (zero) which means that formants will be computed up to the end of the file.}

\item{windowShift}{the analysis window shift length (in ms).}

\item{numFormants}{the number of formants that the analysis should try to find}

\item{windowSize}{the analysis window length (in ms).}

\item{preemphasis}{the frequency from which a preemphasis will be applied.}

\item{ceilingStepSize}{The function multiple searches for formant tracks with the frequency ceiling set to \code{maxhzformant*exp(-ceilingStepSize*stepsUpDown)} to \code{maxhzformant*exp(ceilingStepSize*stepsUpDown)}.}

\item{stepsUpDown}{The number of iterations of increases and decreases of the frequency ceiling to use when trying to find the an optimal formant track.}

\item{windowShape}{the analysis window function used when extracting part of a sound file for analysis. De faults to "Hanning".}

\item{relativeWidth}{the relative width of the windowing function used.}

\item{spectWindowShape}{The shape of the windowing function used for constructing the spectrogram.}

\item{spectResolution}{The frequency resolution of the spectrogram from which formant intensities will be collected.}

\item{toFile}{write the output to a file? The file will be written in  \code{outputDirectory}, if defined, or in the same directory as the soundfile.}

\item{explicitExt}{the file extension that should be used.}

\item{outputDirectory}{set an explicit directory for where the signal file will be written. If not defined, the file will be written to the same directory as the sound file.}

\item{verbose}{Not implemented. Only included here for compatibility.}

\item{praat_path}{give an explicit path for Praat. If the praat}

\item{maxhzformant}{The middle of the formant frequency ceilings that will be attempted when searching for an optimal track.}
}
\value{
The number of processed files, or an SSFF track data object (if \code{toFile=FALSE}) containing three fields ("F", "B" and "L") containing formant frequencies, bandwidth and intensities.
}
\description{
This function exposes Praat's functionality for iteratively searching for the best fit formant track for a file by adjusting the maximum formant frequency (frequency ceiling). In each iteration, Praat's built in function (burg algorithm). See \insertCite{Escudero.2009.10.1121/1.3180321} for example of how the procedure has been used and \inserCite{Weenink} for a description of how the optimal formant track is identified.
If \code{stepsUpDown} is zero, then this function and the \code{praat_formant_burg} function would produce the same result if identical settings are used.
The function also computes the intensity (L) of the best fit formant tracks based on the power of the spectrum at the frequency of the formant. Naturally, if the algorithm failed to find a formant in a specified time frame, then the function will not return a formant frequency, bandwidth and intensity estimation.
}
\details{
If the user only want to estimate formant frequencies that should later be manually corrected, computing them using the function \link[wrassp:forest]{wrassp::forest} or even \code{praat_formant_burg} is \emph{much} quicker. The user should consider this function only if the use case specifically demands an iterative serch for best fit formants.
}
\seealso{
praat_formant_burg
}
