% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat_ssff.R
\name{praat_intensity}
\alias{praat_intensity}
\title{Compute a sound signal intensity track using Praat}
\usage{
praat_intensity(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 80,
  subtractMean = TRUE,
  window = "Gaussian1",
  relativeWidth = 1,
  toFile = TRUE,
  explicitExt = "int",
  outputDirectory = NULL,
  verbose = TRUE,
  praat_path = NULL
)
}
\arguments{
\item{listOfFiles}{A list of sound files that should be analyzed.}

\item{beginTime}{The time (in s) where the analysis should begin.}

\item{endTime}{The time (in s) where the analysis should end}

\item{windowShift}{The time step (in ms) of the resulting intensity contour.}

\item{minF}{The minimum periodicity of the sound signal. If set too
high, the intensity variations within a pitch period will influence the
computed intensity contour. If set too low, the smearing of the intensity
contour may hide rapid intensity variations.}

\item{subtractMean}{Should the average intensity be subtracted in order to
compensate for the constant pressure of the recording environment?}

\item{window}{Which windowing function should be applied when extracting part
of the recording for analysis? Allowed windowing functions are
"rectangular", "triangular", "parabolic", "Hanning", "Hamming",
"Gaussian1", "Gaussian2", "Gaussian3", "Gaussian4", "Gaussian5", "Kaiser1",
and "Kaiser2". Consult the Praat manual for more details.}

\item{relativeWidth}{The relative with of the window used when extracting
part of a sound file for analysis.}

\item{toFile}{Should the SSFF track be written to file (TRUE) or returned as
an object (FALSE)}

\item{explicitExt}{The SSFF file written to disk will have the same name as
the original sound file, but with this file extension.}

\item{outputDirectory}{The directory where the SSFF track will be stored.
Defaults to the same directory as the sound file.}

\item{verbose}{For comparability with wrassp functions, and expected by EmuR. Nothing happens if you set it to FALSE:}

\item{praat_path}{An explicit path to the Praat executable.}
}
\value{
An SSFF object containing the intensity track (if \code{toFile==FALSE}).
}
\description{
This function calls Praat to compute an intensity contour for a sound file,
and return the result as a Assp Data Object that is stored in an SSFF file
(the default), or returned as an object. The analysis will attempt to
compensate for the effect of the periodicity introduced by pitch of the
produced speech, as well as the constant pressure in the recording
environment if supplied with appropriate arguments.
}
