% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_python_pm_pvoice_tremor.R
\name{praat_voice_tremor}
\alias{praat_voice_tremor}
\title{Voice Tremor Analysis using Parselmouth (memory-based, optimized)}
\description{
Voice Tremor Analysis using Parselmouth (memory-based, optimized)
}
