% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_core.R
\name{print.JsonTrackObj}
\alias{print.JsonTrackObj}
\title{Print method for JsonTrackObj}
\usage{
\method{print}{JsonTrackObj}(x, ...)
}
\arguments{
\item{x}{JsonTrackObj}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for JsonTrackObj
}
