% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_python_sptk_rapt.R
\name{process_rapt_single}
\alias{process_rapt_single}
\title{Internal function to process single file with RAPT}
\usage{
process_rapt_single(
  soundFile,
  beginTime,
  endTime,
  windowShift,
  minF,
  maxF,
  voicing.threshold,
  explicitExt,
  outputDirectory,
  toFile
)
}
\description{
Internal function to process single file with RAPT
}
\keyword{internal}
