% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trk_formants_tvwlp.R
\name{process_with_av_python}
\alias{process_with_av_python}
\title{Helper function to process audio with Python backend}
\usage{
process_with_av_python(
  listOfFiles,
  python_func,
  python_kwargs,
  output_track_names,
  output_format,
  beginTime,
  endTime,
  explicitExt,
  outputDirectory,
  toFile,
  verbose,
  track_description
)
}
\description{
Helper function to process audio with Python backend
}
\keyword{internal}
