% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_python_pyin.R
\name{pyin}
\alias{pyin}
\title{Estimate pitch using the probabilistic YIN algorithm}
\usage{
pyin(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  windowSize = 30,
  minF = 70,
  maxF = 200,
  max_transition_rate = 35.92,
  beta_parameters = c(2, 18),
  center = TRUE,
  boltzmann_parameter = 2,
  resolution = 0.1,
  thresholds = 100,
  switch_probability = 0.01,
  no_trough_probability = 0.01,
  pad_mode = "constant",
  explicitExt = "pyp",
  outputDirectory = NULL,
  toFile = TRUE
)
}
\arguments{
\item{listOfFiles}{Vector of file paths to process}

\item{beginTime}{Start time in seconds (default: 0.0)}

\item{endTime}{End time in seconds (default: 0.0 = end of file)}

\item{windowShift}{Frame shift in milliseconds (default: 10.0)}

\item{minF}{Minimum F0 in Hz (default: 60.0)}

\item{maxF}{Maximum F0 in Hz (default: 400.0)}

\item{max_transition_rate}{The maximum pitch transition rate in octaves per second.}

\item{beta_parameters}{The shape parameters for the beta distribution prior over thresholds.}

\item{center}{Should analysis windows be centered around the time of the
window (\code{TRUE}, the default) or should the window be considered to have
started at the indicated time point (\code{FALSE}).}

\item{boltzmann_parameter}{The shape parameter for the Boltzmann distribution prior over troughs. Larger values will assign more mass to smaller periods.}

\item{resolution}{The resolution of the pitch bins. 0.01 corresponds to cents.}

\item{thresholds}{The number of thresholds for peak estimation.}

\item{switch_probability}{The probability of switching from voiced to unvoiced or vice versa.}

\item{no_trough_probability}{The maximum probability to add to global minimum if no trough is below threshold.}

\item{pad_mode}{The mode in which padding occurs. Ignored if \code{center} is not
\code{TRUE}. Padding occurs in the python library librosa, and the user should
therefore consult the manual of the NumPy library function
\href{https://numpy.org/doc/stable/reference/generated/numpy.pad.html#numpy-pad}{numpy.pad} for other options.}

\item{explicitExt}{Output file extension (default: "f0")}

\item{outputDirectory}{Output directory (default: NULL = same as input)}

\item{toFile}{Write results to file (default: TRUE)}
}
\value{
An SSFF track object containing two tracks (f0 and pitch) that are either returned (toFile == FALSE) or stored on disk.
}
\description{
The probabilistic YIN algorithm
\insertCite{Mauch.2014.10.1109/icassp.2014.6853678}{superassp} is an
extension of \link[=yin]{YIN} \insertCite{Cheveigné.2002.10.1121/1.1458024}{superassp} that considers multiple pitch
candidates in a hidden Markov model that is Viterbi-decoded to deduce the
final pitch estimate. The function also returns a track encoding whether the
track was considered voiced or not, and a track containing the probability of
voicing in the analysis frame.
}
\details{
This function calls the librosa \insertCite{brian_mcfee_2022_6097378}{superassp} Python library to load the audio data an
make pitch related estimates.
}
\references{
\insertAllCited{}
}
