% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_ssff.R
\name{rapt}
\alias{rapt}
\title{Compute f0 using the RAPT algorithm (Optimized)}
\usage{
rapt(
  listOfFiles = NULL,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 70,
  maxF = 200,
  voicing.threshold = 0.3,
  explicitExt = "rpt",
  outputDirectory = NULL,
  toFile = TRUE,
  assertLossless = NULL,
  logToFile = FALSE,
  keepConverted = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{listOfFiles}{A vector of file paths to wav files.}

\item{beginTime}{The start time of the section of the sound file that should be processed.}

\item{endTime}{The end time of the section of the sound file that should be processed.}

\item{windowShift}{The measurement interval (frame duration), in seconds.}

\item{minF}{Candidate f0 frequencies below this frequency will not be considered.}

\item{maxF}{Candidates above this frequency will be ignored.}

\item{voicing.threshold}{Voice/unvoiced threshold. Default is 0.3.}

\item{explicitExt}{File extension for output files}

\item{outputDirectory}{Output directory path (NULL for same as input)}

\item{toFile}{Write output to file (TRUE) or return object (FALSE)}

\item{assertLossless}{an optional list of file extensions that the user wants to assert
contains losslessly encoded signals data.}

\item{logToFile}{whether to log commands to a separate logfile in the
\code{outputDirectory}. Logging will otherwise be in the function-specific logging
namespace of \link[logger:logger]{logger::logger} and will be put wherever this namespace is defined to place its output.
See \link[logger:log_appender]{logger::log_appender} for details.}

\item{keepConverted}{whether to keep converted files}

\item{verbose}{display verbose information about processing steps taken, as
well as progress bars.}
}
\value{
If \code{toFile} is \code{FALSE}, the function returns a list of AsspDataObj
objects. If \code{toFile} is \code{TRUE}, the number (integer) of successfully
processed and stored output files is returned.
}
\description{
Optimized version of RAPT f0 computation following the forest function template.
}
