% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_io.R
\name{read_json_track_jsonlite}
\alias{read_json_track_jsonlite}
\title{Read JSON using jsonlite (fallback)}
\usage{
read_json_track_jsonlite(file)
}
\arguments{
\item{file}{Path to JSON file}
}
\value{
JsonTrackObj
}
\description{
Read JSON using jsonlite (fallback)
}
\keyword{internal}
