% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_io.R
\name{read_json_track_simdjson}
\alias{read_json_track_simdjson}
\title{Read JSON using RcppSimdJson (fast)}
\usage{
read_json_track_simdjson(file)
}
\arguments{
\item{file}{Path to JSON file}
}
\value{
JsonTrackObj
}
\description{
Read JSON using RcppSimdJson (fast)
}
\keyword{internal}
