% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_helpers.R
\name{rmsana_memory}
\alias{rmsana_memory}
\title{Perform RMS analysis on AsspDataObj in memory}
\usage{
rmsana_memory(audio_obj, ...)
}
\arguments{
\item{audio_obj}{AsspDataObj containing audio data}

\item{...}{Additional parameters passed to rmsana}
}
\value{
AsspDataObj with RMS analysis results
}
\description{
This function performs RMS analysis on an AsspDataObj without writing
intermediate files. It's useful when processing audio data read from
video files or other non-WAV formats using the av package.
}
\examples{
\dontrun{
audio_obj <- av_to_asspDataObj("myfile.mp4")
rms_result <- rmsana_memory(audio_obj, windowShift = 5)
}
}
