% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtorch_deepformants.R
\name{run_deepformants_estimator}
\alias{run_deepformants_estimator}
\title{Run DeepFormants MLP estimator on single feature vector}
\usage{
run_deepformants_estimator(feat_mat)
}
\arguments{
\item{feat_mat}{(1, 350) numeric matrix}
}
\value{
length-4 numeric vector of formant frequencies in Hz
}
\description{
Run DeepFormants MLP estimator on single feature vector
}
\keyword{internal}
