% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtorch_deepformants.R
\name{run_deepformants_tracker}
\alias{run_deepformants_tracker}
\title{Run DeepFormants LSTM tracker on feature matrix}
\usage{
run_deepformants_tracker(feat_mat)
}
\arguments{
\item{feat_mat}{(n_frames, 350) numeric matrix}
}
\value{
(n_frames, 4) matrix of formant frequencies in Hz
}
\description{
Run DeepFormants LSTM tracker on feature matrix
}
\keyword{internal}
