% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_sacc.R
\name{sacc_available}
\alias{sacc_available}
\title{Check if SAcC pitch tracker is available}
\usage{
sacc_available()
}
\value{
Logical. TRUE if all required modules are available, FALSE otherwise.
}
\description{
Checks whether the required Python modules for SAcC are available.
}
\examples{
\dontrun{
if (sacc_available()) {
  result <- trk_sacc("speech.wav", toFile = FALSE)
} else {
  install_sacc()
}
}

}
\seealso{
\code{\link{install_sacc}} to install SAcC dependencies
\code{\link{sacc_info}} to get SAcC configuration information
}
