% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_sacc.R
\name{sacc_info}
\alias{sacc_info}
\title{Get SAcC pitch tracker information}
\usage{
sacc_info()
}
\value{
A list with information about:
\itemize{
\item Python path and version
\item NumPy version
\item SciPy version
\item soundfile availability
\item SAcC module path
\item Available configuration files
}
}
\description{
Returns information about the SAcC installation and configuration.
}
\examples{
\dontrun{
if (sacc_available()) {
  info <- sacc_info()
  print(info)
}
}

}
\seealso{
\code{\link{install_sacc}} to install SAcC dependencies
\code{\link{sacc_available}} to check availability
}
