% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat_internal_se_vq.R
\name{search_res_interval_peaks}
\alias{search_res_interval_peaks}
\title{Find top-N residual peaks in each MBS interval}
\usage{
search_res_interval_peaks(res, interval, Ncand)
}
\description{
Find top-N residual peaks in each MBS interval
}
\keyword{internal}
