% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_python_seenc.R
\name{seenc}
\alias{seenc}
\title{(En)coded spectral envelope estimation}
\usage{
seenc(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 70,
  maxF = 200,
  dimensions = 1,
  explicitExt = "sec",
  outputDirectory = NULL,
  toFile = TRUE
)
}
\arguments{
\item{listOfFiles}{Vector of file paths to process}

\item{beginTime}{Start time in seconds (default: 0.0)}

\item{endTime}{End time in seconds (default: 0.0 = end of file)}

\item{windowShift}{Frame shift in milliseconds (default: 10.0)}

\item{minF}{Minimum F0 in Hz (default: 60.0)}

\item{maxF}{Maximum F0 in Hz (default: 400.0)}

\item{dimensions}{Number of dimensions of coded spectral envelope}

\item{explicitExt}{Output file extension (default: "f0")}

\item{outputDirectory}{Output directory (default: NULL = same as input)}

\item{toFile}{Write results to file (default: TRUE)}
}
\value{
An SSFF track object containing two tracks (f0 and corr) that are
either returned (toFile == FALSE) or stored on disk.
}
\description{
The Spectral envelope is estimated using function for assessing band-aperiodicities
using the CheapTrick algorithm
\insertCite{Morise:2015ia}{superassp} implemented
in the WORLD vocoder
\insertCite{MORISE.2016.10.1587/transinf.2015edp7457}{superassp}. The the \link{harvest}
pitch algorithm \insertCite{Morise.2017.10.21437/interspeech.2017-68}{superassp} is used to calculate
the periodic component.
}
\references{
\insertAllCited{}
}
