% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptk_helpers.R, R/utils_av_sptk_helpers.R
\name{sptk_helpers}
\alias{sptk_helpers}
\title{SPTK Helper Functions}
\description{
Internal helper functions shared across SPTK-based DSP functions

Internal helper functions shared across SPTK-based DSP functions
}
\keyword{internal}
