% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat_ssff.R
\name{ssffToFormant}
\alias{ssffToFormant}
\title{Make a Praat Formant object from an SSFF track track}
\usage{
ssffToFormant(
  inData,
  outputPath = NULL,
  fm.field = "fm",
  bw.field = "bw",
  start = 0,
  end = 0,
  windowShift = 5,
  nominalF1 = 500,
  zero.threshold = 0,
  dump.script = FALSE,
  soundFileDuration = NULL
)
}
\arguments{
\item{inData}{An SSFF formant track object, or the full path to one.}

\item{outputPath}{The directory where the Formant file should be stored.}

\item{fm.field}{The field / column in the SSFF object where the formant frequency tracks
are stored. The field may be indicated by the name of it, or its number.
Often, the SSFF object will contain formant frequency values in the first field and bandwidths in the second,
so the user could also give \code{field=1} as an argument.}

\item{bw.field}{The field / column in the SSFF object where the formant banwidth tracks
are stored. The field may be indicated by the name of it, or its number.
Often, the SSFF object will contain formant frequency values in the first field and bandwidths in the second,
so the user could also give \code{field=2} as an argument.}

\item{start}{An optional start time (in s) for a part of the SSFF formant track
which should be converted to a Formant object.}

\item{end}{An optional end time (in s) for a part of the SSFF formant track which
should be converted to a Formant object.}

\item{windowShift}{The analysis window of the formant track (in ms).}

\item{nominalF1}{An assumed F$_1$ value for the created Formant object.}

\item{zero.threshold}{The threshold below which formant frequencies or bandwidths should be considered as zero.}

\item{dump.script}{If \code{TRUE}, the Praat script that is used to create a Pitch object will be dumped to a file for debuging or inspection. The file will be placed in the same directory as the output Formant file, but with a '.praat'. extension.}

\item{soundFileDuration}{An explicit duration that will be set for the created
Formant object. If not explicitly supplied, the duration of the SSFF object track will be used instead.}
}
\description{
This function takes an SSFF object or a path to a file containing one, reads
the content of a \code{fm.field} and \code{bw.field} and uses
Praat to construct a Formant object and store it in a file. The function's primary purpose is to allow the user to re-use already computed, and possibly adjusted, formant frequency and bandwidth tracks in subsequent calculations in for instance \link{praat_sauce}.
}
\details{
The user may optionally compute a Formant object for only parts of the SSFF formant track if the track is stored in a file. \emph{An ability to subset an SSFF object directly is currently not implemented}.

Please note that the process of converting formant tracks from SSFF is \emph{very} slow, as shown by the
\link{microbenchmark} output below which indicates that constructing an Formant object using the current strategy takes about
353 times the time it takes to simply compute the formant tracks.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Unit: relative
expr         min          lq        mean      median          uq         max neval
read      1.0000      1.0000      1.0000      1.0000      1.0000      1.0000     1
forest    705.8529    705.8529    705.8529    705.8529    705.8529    705.8529     1
Formant 249267.5829 249267.5829 249267.5829 249267.5829 249267.5829 249267.5829     1
}\if{html}{\out{</div>}}
}
\seealso{
\link{praat_sauce}

\link[wrassp:forest]{wrassp::forest}

\link{praat_formant_burg}
}
