% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_methods.R
\name{subset_json_track}
\alias{subset_json_track}
\title{Subset JsonTrackObj}
\usage{
subset_json_track(x, start_time = NULL, end_time = NULL)
}
\arguments{
\item{x}{JsonTrackObj}

\item{start_time}{Start time in seconds}

\item{end_time}{End time in seconds}
}
\value{
JsonTrackObj with filtered slices
}
\description{
Extract slices within a time range
}
