% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_avaudio.R
\name{summary.AVAudio}
\alias{summary.AVAudio}
\title{Summary Method for AVAudio}
\arguments{
\item{object}{AVAudio object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for AVAudio
}
\keyword{internal}
