% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_methods.R
\name{summary.JsonTrackObj}
\alias{summary.JsonTrackObj}
\title{Summary method for JsonTrackObj}
\usage{
\method{summary}{JsonTrackObj}(object, ...)
}
\arguments{
\item{object}{JsonTrackObj}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for JsonTrackObj
}
