% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_voice_analysis.R
\name{swiftf0_info}
\alias{swiftf0_info}
\title{Get Swift-F0 module information}
\usage{
swiftf0_info()
}
\value{
A list with module information and model specifications
}
\description{
Get Swift-F0 module information
}
\examples{
\dontrun{
info <- swiftf0_info()
print(info)
}
}
