% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_ssff.R
\name{torch_pitch}
\alias{torch_pitch}
\title{Pitch tracking using the torch pitch tracker}
\usage{
torch_pitch(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 10,
  windowSize = 30,
  minF = 70,
  maxF = 200,
  explicitExt = "tpi",
  outputDirectory = NULL,
  toFile = TRUE
)
}
\arguments{
\item{listOfFiles}{A vector of file paths to wav files.}

\item{beginTime}{The start time of the section of the sound file that should be processed.}

\item{endTime}{The end time of the section of the sound file that should be processed.}

\item{windowShift}{The measurement interval (frame duration), in seconds.}

\item{minF}{Candidate f0 frequencies below this frequency will not be considered.}

\item{maxF}{Candidates above this frequency will be ignored.}

\item{explicitExt}{File extension for output files}

\item{outputDirectory}{Output directory path (NULL for same as input)}

\item{toFile}{Write output to file (TRUE) or return object (FALSE)}
}
\value{
An SSFF track object containing two tracks (f0 and pitch) that are either returned (toFile == FALSE) or stored on disk.
}
\description{
This function estimates pitch by normalized cross-correlation function (NCCF) and
median smoothing, as implemented in the torchaudio
\insertCite{yang2021torchaudio}{superassp} library. The exact algorithm is undisclosed by the implementing library
but approach likely builds on earlier implementations that use NCCFs \insertCite{talkin1995robust,Kasi.2002.10.1109/icassp.2002.5743729}{superassp}
including the \link[superassp::rapt]{RAPT} algorithm.
}
\references{
\insertAllCited{}
}
\seealso{
rapt
}
