% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_python_aperiodicities.R
\name{trk_aperiodicities}
\alias{trk_aperiodicities}
\title{Estimates aperiodicity of a speech signal (DEPRECATED)}
\usage{
trk_aperiodicities(listOfFiles, ...)
}
\arguments{
\item{listOfFiles}{Vector of file paths to process}
}
\value{
An SSFF track object containing two tracks (f0 and corr) that are
either returned (toFile == FALSE) or stored on disk.
}
\description{
\strong{DEPRECATED}: This function is deprecated. Please use \code{\link[=trk_d4c]{trk_d4c()}} instead,
which provides a faster C++ implementation with the same functionality.

Aperiodicity is estimated using function for assessing band-aperiodicities
using the Definitive Decomposition Derived Dirt-Cheap (D4C) algorithm
\insertCite{Morise.2016.10.1016/j.specom.2016.09.001}{superassp} implemented
in the WORLD vocoder
\insertCite{MORISE.2016.10.1587/transinf.2015edp7457}{superassp}. The the \code{dio()}
\insertCite{morise2010rapid}{superassp} pitch algorithm is used to calculate
the periodic component.
}
\references{
\insertAllCited{}
}
