% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_python_crepe.R
\name{trk_crepe}
\alias{trk_crepe}
\title{Compute pitch and periodicity using the CREPE pitch tracker}
\usage{
trk_crepe(listOfFiles, ...)
}
\arguments{
\item{listOfFiles}{A vector of file paths to wav files.}

\item{beginTime}{(Not implemented) The start time of the section of the sound file that should be processed.}

\item{endTime}{(Not implemented) The end time of the section of the sound file that should be processed.}

\item{windowShift}{The measurement interval (frame duration), in seconds.}

\item{minF}{Candidate f0 frequencies below this frequency will not be considered.}

\item{maxF}{Candidates above this frequency will be ignored.}

\item{voicing.threshold}{Voice/unvoiced threshold. Default is 0.21.}

\item{silence.threshold}{Frames that do not contain amplitudes above this threshold (relative to the global maximum amplitude), are probably silent.}

\item{model}{Use a fast ("tiny") model, or a more complete ("full") model to find pitch. The more complete model will take approximately 9-11 times longer to process the file.}

\item{conda.env}{The name of the conda environment in which Python and its
required packages are stored. Please make sure that you know what you are
doing if you change this. Defaults to \code{NULL}, which means that the default enviroment or the environment set in the
\code{RETICULATE_PYTHON} environment variable will be used.}
}
\value{
An SSFF track object containing two tracks (f0 and periodicity) that are either returned (toFile == FALSE) or stored on disk.
}
\description{
The CREPE \insertCite{Kim.2018.10.1109/icassp.2018.8461329}{superassp} applies a deep convolutional neural network directly on the time-domain waveform to find
the fundamental frequency in a speech signal. Two versions of the models have been trained, one smaller yielding quicker results, and the full model which can be considerably
more computationally intensive to apply.
}
\details{
T
}
\references{
\insertAllCited{}
}
