% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_c_assp_cssSpectrum.R
\name{trk_cssSpectrum}
\alias{trk_cssSpectrum}
\title{Compute the cepstrally smoothed spectrum (Rcpp-optimized)}
\usage{
trk_cssSpectrum(
  listOfFiles = NULL,
  beginTime = 0,
  centerTime = FALSE,
  endTime = 0,
  resolution = 40,
  fftLength = 0,
  windowShift = 5,
  numCeps = 0,
  window = "BLACKMAN",
  toFile = TRUE,
  explicitExt = "css",
  outputDirectory = NULL,
  assertLossless = NULL,
  logToFile = FALSE,
  keepConverted = FALSE,
  convertOverwrites = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{numCeps}{= \if{html}{\out{<num>}}: set number of cepstral coefficients used to \if{html}{\out{<num>}}
(default: sampling rate in kHz + 1; minimum: 2)}
}
\value{
The number of successfully written files (if \code{toFile=TRUE}), or a vector of \code{AsspDataObj} objects (if \code{toFile=FALSE}).
}
\description{
Short-term spectral analysis of the signals in  \code{listOfFiles}
using the Fast Fourier Transform and cepstral smoothing.
}
\details{
The results will be will be written to an SSFF formated file with the base
name of the input file and extension \emph{.css} in a track \emph{CSS\link{dB}} which contains amplitudes (on a dB scale) of
all frequencies in the computed spectrum.

The function is a re-write of the \link[wrassp:cssSpectrum]{wrassp::cssSpectrum} function, but
with media pre-conversion, better checking of preconditions such as the
input file existence, structured logging, and the use of a more modern
framework for user feedback. This version includes Rcpp optimizations
for improved performance on large batches of files.

The native file type of this function is "wav" files (in "pcm_s16le"
format), SUNs "au", NIST, or CSL formats (kay or NSP extension). Input
signal conversion, when needed, is done by
\href{https://ffmpeg.org/libavcodec.html}{libavcodec} and the excellent \link[av:encoding]{av::av_audio_convert}
wrapper function
}
\note{
This function takes some time to apply but also result in data in a relatively large matrix.
It is therefore not usually efficient to store intermediate results in a cache.
However, if the number of signals it will be applied to
is \emph{very} large, then caching of results may be warranted.
}
\examples{
# get path to audio file
path2wav <- list.files(system.file("samples","sustained", package = "superassp"), pattern = glob2rx("a1.wav"), full.names = TRUE)

# calculate cepstrally smoothed spectrum
res <- trk_cssSpectrum(path2wav, toFile=FALSE)
resolution <- attr(res,"origFreq") / ncol(res[[1]])

# plot spectral values at midpoint of signal
plot(y=res[["CSS[dB]"]][400,],
    x=seq(1,ncol(res[[1]]),1)* resolution,
    type='l',
    xlab='Frequency (Hz)',
    ylab='Amplitude (dB)')

}
\seealso{
\code{\link{dftSpectrum}}, \code{\link{lpsSpectrum}}, \code{\link{cepstrum}};
all derived from libassp's spectrum function.

\link[wrassp:cssSpectrum]{wrassp::cssSpectrum}

\link{AsspWindowTypes}

\link[av:encoding]{av::av_audio_convert}
}
\author{
Raphael Winkelmann

Lasse Bombien

Fredrik Nylén
}
