% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_c_assp_dftSpectrum.R
\name{trk_dftSpectrum}
\alias{trk_dftSpectrum}
\title{Computes the Discrete Fourier Transform spectrum (Rcpp-optimized)}
\usage{
trk_dftSpectrum(
  listOfFiles = NULL,
  beginTime = 0,
  centerTime = FALSE,
  endTime = 0,
  resolution = 40,
  fftLength = 0,
  windowShift = 5,
  window = "BLACKMAN",
  bandwidth = 0,
  toFile = TRUE,
  explicitExt = "dft",
  outputDirectory = NULL,
  assertLossless = NULL,
  logToFile = FALSE,
  keepConverted = FALSE,
  convertOverwrites = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{listOfFiles}{vector of file paths to be processed by function}

\item{resolution}{= \if{html}{\out{<freq>}}: set FFT length to the smallest value which
results in a frequency resolution of \if{html}{\out{<freq>}} Hz or better (default: 40.0)}

\item{fftLength}{= \if{html}{\out{<num>}}: set FFT length to \if{html}{\out{<num>}} points (overrules default
and 'resolution' option)}

\item{bandwidth}{= \if{html}{\out{<freq>}}: set the effective analysis bandwidth to \if{html}{\out{<freq>}}
Hz (default: 0, yielding the smallest possible value given the length of
the FFT)}
}
\value{
The number of successfully written files (if \code{toFile=TRUE}), or a vector of \code{AsspDataObj} objects (if \code{toFile=FALSE}).
}
\description{
Short-term spectral analysis of the signal in \if{html}{\out{<listOfFiles>}}
using the Fast Fourier Transform. The default is to
calculate an unsmoothed narrow-band spectrum with the
size of the analysis window equal to the length of the
FFT. The output from the FFT will be converted to a
power spectrum in dB from 0 Hz up to and including the
Nyquist rate.
}
\details{
The results will be will be written to an SSFF formated file with the base
name of the input file and extension \emph{.dft} in a track \emph{DFT[dB]} which contains amplitudes (on a dB scale) of
all frequencies in the computed spectrum.

The function is a re-write of the \link[wrassp:dftSpectrum]{wrassp::dftSpectrum} function, but
with media pre-conversion, better checking of preconditions such as the
input file existence, structured logging, and the use of a more modern
framework for user feedback. The \emph{libassp} \insertCite{s5h}{superassp} C library code is used for
DSP. This version includes Rcpp optimizations for improved performance on large batches of files.

The native file type of this function is "wav" files (in "pcm_s16le"
format), SUNs "au", NIST, or CSL formats (kay or NSP extension). Input
signal conversion, when needed, is done by
\href{https://ffmpeg.org/libavcodec.html}{libavcodec} and the excellent \link[av:encoding]{av::av_audio_convert}
wrapper function
}
\note{
This function takes some time to apply but also result in data in a relatively large matrix.
It is therefore not usually efficient to store intermediate results in a cache.
However, if the number of signals it will be applied to
is \emph{very} large, then caching of results may be warranted.
}
\examples{
# get path to audio file
path2wav <- list.files(system.file("extdata", package = "wrassp"),
                       pattern = glob2rx("*.wav"),
                       full.names = TRUE)[1]

# calculate dft spectrum
res <- trk_dftSpectrum(path2wav, toFile=FALSE)

# plot spectral values at midpoint of signal
plot(res$dft[dim(res$dft)[1]/2,],
     type='l',
     xlab='spectral value index',
     ylab='spectral value')

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{cssSpectrum}}, \code{\link{lpsSpectrum}}, \code{\link{cepstrum}};
all derived from \emph{libassp} \insertCite{s5h}{superassp} spectrum function.
}
\author{
Raphael Winkelmann

Lasse Bombien

Fredrik Nylén
}
