% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_python_pm_pintensity.R
\name{trk_intensityp}
\alias{trk_intensityp}
\title{Optimized intensity analysis using Python/Parselmouth}
\usage{
trk_intensityp(listOfFiles, ...)
}
\arguments{
\item{listOfFiles}{Vector of file paths to audio files}

\item{beginTime}{Start time in seconds (0 for beginning of file)}

\item{endTime}{End time in seconds (0 for end of file)}

\item{time_step}{Time step between frames in seconds (0 for automatic)}

\item{minimal_f0_frequency}{Minimum pitch frequency in Hz}

\item{subtract_mean}{Whether to subtract the mean intensity}

\item{windowShape}{Window shape for time windowing}

\item{relativeWidth}{Relative width for windowing}

\item{toFile}{Write output to file (TRUE) or return object (FALSE)}

\item{explicitExt}{File extension for output files}

\item{outputDirectory}{Output directory path (NULL for same as input)}
}
\value{
If toFile=TRUE, returns number of files successfully processed.
If toFile=FALSE, returns an AsspDataObj with intensity track.
}
\description{
This is an optimized version of intensity analysis that uses Python's
Parselmouth library instead of calling external Praat. It computes the
intensity (loudness) contour of a sound signal.
}
