% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_c_assp_lpsSpectrum.R
\name{trk_lpsSpectrum}
\alias{trk_lpsSpectrum}
\title{Calculate Linear Prediction smoothed spectrum (Rcpp-optimized)}
\usage{
trk_lpsSpectrum(listOfFiles, ...)
}
\arguments{
\item{resolution}{= \if{html}{\out{<freq>}}: set FFT length to the smallest value which
results in a frequency resolution of \if{html}{\out{<freq>}} Hz or better (default: 40.0)}

\item{fftLength}{= \if{html}{\out{<num>}}: set FFT length to \if{html}{\out{<num>}} points (overrules default
and 'resolution' option)}

\item{order}{= \if{html}{\out{<num>}}: set prediction order to \if{html}{\out{<num>}} (default: sampling
rate in kHz + 3)}

\item{preemphasis}{= \if{html}{\out{<val>}}: set pre-emphasis factor to \if{html}{\out{<val>}} (default:
-0.95)}

\item{deemphasize}{(default: undo spectral tilt due to
pre-emphasis used in LP analysis, i.e. TRUE)}

\item{windowSize}{window size in milliseconds}
}
\value{
The number of successfully written files (if \code{toFile=TRUE}), or a vector of \code{AsspDataObj} objects (if \code{toFile=FALSE}).
}
\description{
Short-term spectral analysis of the signal in \code{listOfFiles}
using the Fast Fourier Transform and linear predictive smoothing.
}
\details{
The results will be will be written to an SSFF formated file with the base
name of the input file and extension \emph{.lps} in a track \emph{LPS\link{dB}} which contains amplitudes (on a dB scale) of
all frequencies in the computed spectrum.

The function is a re-write of the \link[wrassp:lpsSpectrum]{wrassp::lpsSpectrum} function, but
with media pre-conversion, better checking of preconditions such as the
input file existence, structured logging, and the use of a more modern
framework for user feedback. This version includes Rcpp optimizations
for improved performance on large batches of files.

The native file type of this function is "wav" files (in "pcm_s16le"
format), SUNs "au", NIST, or CSL formats (kay or NSP extension). Input
signal conversion, when needed, is done by
\href{https://ffmpeg.org/libavcodec.html}{libavcodec} and the excellent \link[av:encoding]{av::av_audio_convert}
wrapper function
}
\note{
This function takes some time to apply but also result in data in a relatively large matrix.
It is therefore not usually efficient to store intermediate results in a cache.
However, if the number of signals it will be applied to
is \emph{very} large, then caching of results may be warranted.
}
\examples{
# get path to audio file
path2wav <- list.files(system.file("samples","sustained", package = "superassp"), pattern = glob2rx("a1.wav"), full.names = TRUE)

# calculate linear prediction smoothed spectrum
res <- trk_lpsSpectrum(path2wav, toFile=FALSE)
resolution <- attr(res,"origFreq") / ncol(res[[1]])

# plot spectral values at midpoint of signal
plot(y=res[["CSS[dB]"]][400,],
    x=seq(1,ncol(res[[1]]),1)* resolution,
    type='l',
    xlab='Frequency (Hz)',
    ylab='Amplitude (dB)')

}
\seealso{
\code{\link{dftSpectrum}}, \code{\link{cssSpectrum}}, \code{\link{cepstrum}};
all derived from \emph{libassp} \insertCite{s5h}{superassp} spectrum function.
}
\author{
Raphael Winkelmann

Lasse Bombien

Fredrik Nylén
}
