% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_c_assp_rmsana.R
\name{trk_rmsana}
\alias{trk_rmsana}
\title{Short-term Root Mean Square amplitude of signals (Rcpp-optimized)}
\usage{
trk_rmsana(listOfFiles, ...)
}
\arguments{
\item{linear}{Should linear RMS values be computed? The default (\code{FALSE})
means that the output will be on a logarithmic decibel scale (dB).}
}
\value{
The number of successfully written files (if \code{toFile=TRUE}), or a vector of \code{AsspDataObj} objects (if \code{toFile=FALSE}).
}
\description{
The RMS amplitude is computed for each window of \code{windowSize}
length in the input signals files listed in \code{listOfFiles}. Per default, the
RMS values are expressed in decibel (dB) so that they correspond to the
short-term power of the signal. Input signals not in a file format natively
supported will be converted before the autocorrelation functions are
computed. The conversion process will display warnings about input files
that are not in known losslessly encoded formats.

The results will be will be written to an SSFF formated file with the base
name of the input file and extension \emph{.rms} in a track \emph{RMS}.
}
\details{
The function is a re-write of the \link[wrassp:rmsana]{wrassp::rmsana} function, but
with media pre-conversion, better checking of preconditions such as the
input file existence, structured logging, and the use of a more modern
framework for user feedback. This version includes Rcpp optimizations
for improved performance on large batches of files.
}
\note{
This function is not considered computationally expensive enough to require caching of
results if applied to many signals. However, if the number of signals it will be applied to
is \emph{very} large, then caching of results may be warranted.
}
