% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_python_seenc.R
\name{trk_seenc}
\alias{trk_seenc}
\title{(En)coded spectral envelope estimation}
\usage{
trk_seenc(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 70,
  maxF = 200,
  dimensions = 1,
  explicitExt = "sec",
  outputDirectory = NULL,
  toFile = TRUE
)
}
\arguments{
\item{dimensions}{Number of dimensions of coded spectral envelope}
}
\value{
An SSFF track object containing two tracks (f0 and corr) that are
either returned (toFile == FALSE) or stored on disk.
}
\description{
The Spectral envelope is estimated using function for assessing band-aperiodicities
using the CheapTrick algorithm
\insertCite{Morise:2015ia}{superassp} implemented
in the WORLD vocoder
\insertCite{MORISE.2016.10.1587/transinf.2015edp7457}{superassp}. The the \link{harvest}
pitch algorithm \insertCite{Morise.2017.10.21437/interspeech.2017-68}{superassp} is used to calculate
the periodic component.
}
\references{
\insertAllCited{}
}
