% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_python_torch_pitch.R
\name{trk_torch_pitch}
\alias{trk_torch_pitch}
\title{Pitch tracking using the torch pitch tracker}
\usage{
trk_torch_pitch(listOfFiles, ...)
}
\value{
An SSFF track object containing two tracks (f0 and pitch) that are either returned (toFile == FALSE) or stored on disk.
}
\description{
This function estimates pitch by normalized cross-correlation function (NCCF) and
median smoothing, as implemented in the torchaudio
\insertCite{yang2021torchaudio}{superassp} library. The exact algorithm is undisclosed by the implementing library
but approach likely builds on earlier implementations that use NCCFs \insertCite{talkin1995robust,Kasi.2002.10.1109/icassp.2002.5743729}{superassp}
including the \link[superassp::rapt]{RAPT} algorithm.
}
\references{
\insertAllCited{}
}
\seealso{
rapt
}
