% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_c_assp_zcrana.R
\name{trk_zcrana}
\alias{trk_zcrana}
\title{Analysis of the averages of the short-term positive and negative zero-crossing rates (Rcpp-optimized)}
\usage{
trk_zcrana(listOfFiles, ...)
}
\arguments{
\item{listOfFiles}{vector of file paths to be processed by function}

\item{windowSize}{window size in milliseconds}
}
\value{
If \code{toFile} is \code{FALSE}, the function returns a list of \code{AsspDataObj}
objects. If \code{toFile} is \code{TRUE}, the number (integer) of successfully
processed and stored output files is returned.
}
\description{
Analysis of the averages of the short-term positive and
negative zero-crossing rates of the signal in \code{listOfFiles} using the \emph{libassp} C library
\insertCite{s5h}{superassp} function. If \code{toFile} is \code{TRUE}, the results will be written to an output fil in the SSFF binary format, with the
same name as the input file, but with an extension \emph{.zcr} and with a track named 'ZCR[Hz]'.
}
\details{
Input signals not in a natively supported file format will be converted
before the autocorrelation functions are computed. The conversion process
will display warnings about input files that are not in known losslessly
encoded formats.

The function is a re-write of the \link[wrassp:zcrana]{wrassp::zcrana} function, but
with media pre-conversion, better checking of preconditions such as the
input file existence, structured logging, and the use of a more modern
framework for user feedback. This version includes Rcpp optimizations
for improved performance on large batches of files.

The native file type of this function is "wav" files (in "pcm_s16le"
format), SUNs "au", NIST, or CSL formats (kay or NSP extension). Input
signal conversion, when needed, is done by
\href{https://ffmpeg.org/libavcodec.html}{libavcodec} and the excellent \link[av:encoding]{av::av}
wrapper package.
}
\note{
This function is not considered computationally expensive enough to require caching of
results if applied to many signals. However, if the number of signals it will be applied to
is \emph{very} long, then caching of results may be warranted.
}
\examples{
# get path to audio file
path2wav <- list.files(system.file("samples","sustained", package = "superassp"), pattern = glob2rx("a1.wav"), full.names = TRUE)

# calculate zcr values
res <- trk_zcrana(path2wav, toFile=FALSE)

# plot zcr values
plot(seq(0,numRecs.AsspDataObj(res) - 1) / rate.AsspDataObj(res) +
      attr(res, 'startTime'),
    res[["ZCR[Hz]"]],
    type='l',
    xlab='time (s)',
    ylab='Zero Crossing Rates (Hz)')

}
\author{
Raphael Winkelmann

Lasse Bombien

Fredrik Nylén
}
