% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assp_library_vars.R
\name{useWrasspLogger}
\alias{useWrasspLogger}
\title{package variable to force the usage of the logger
set to FALSE by default}
\usage{
useWrasspLogger
}
\description{
package variable to force the usage of the logger
set to FALSE by default
}
\author{
Raphael Winkelmann
}
\keyword{internal}
