% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstf_core.R
\name{validate_json_track}
\alias{validate_json_track}
\title{Validate JsonTrackObj}
\usage{
validate_json_track(obj)
}
\arguments{
\item{obj}{A \code{JsonTrackObj} to validate.}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Validate JsonTrackObj
}
