% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_vat.R
\name{vat_available}
\alias{vat_available}
\title{Check Voice Analysis Toolkit Availability}
\usage{
vat_available()
}
\value{
Logical; TRUE if VAT is available, FALSE otherwise
}
\description{
Check if the Voice Analysis Toolkit Python module is available.
}
\examples{
if (vat_available()) {
  # Use VAT functions
} else {
  install_vat()
}

}
