% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat_internal_dsp.R
\name{vat_lpc}
\alias{vat_lpc}
\alias{.vat_lpcauto}
\alias{.vat_burg}
\alias{.vat_lpcar2rf}
\alias{.vat_lpcar2ra}
\alias{.vat_lpcrf2rr}
\alias{.vat_distitar}
\title{LPC primitives}
\usage{
.vat_lpcauto(s, p)

.vat_burg(s, p)

.vat_lpcar2rf(ar)

.vat_lpcar2ra(ar)

.vat_lpcrf2rr(k)

.vat_distitar(ar1, ar2, symmetric = FALSE)
}
\arguments{
\item{s}{input signal}

\item{p}{model order}

\item{ar}{AR coefficients (ar(0) = 1)}

\item{k}{reflection coefficients}

\item{ar1, ar2}{AR coefficient vectors}

\item{symmetric}{symmetric Itakura distance if TRUE}
}
\description{
LPC primitives
}
