% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_ftrack_tvwlp.R
\name{verify_cython_extensions}
\alias{verify_cython_extensions}
\title{Verify Cython extensions were built correctly}
\usage{
verify_cython_extensions(ftrack_path, verbose = TRUE)
}
\description{
Verify Cython extensions were built correctly
}
\keyword{internal}
