% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_slicefunctions.R
\name{voice_analysis}
\alias{voice_analysis}
\title{Applies the Voice Analysis Toolbox on all vowel samples in a directory}
\usage{
voice_analysis(directory)
}
\arguments{
\item{directory}{The directory where sustained vowel samples are stored. The
directory needs to be writable by the user, since the compiled code will
also store the results of computations there.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with one row for each sustained vowel sample,
and with 340 columns. The first column (\code{listOfFiles}) contains the file
names of recordings. The following columns (2 to 340) then contain the
measurement values for the recording.
}
\description{
The Voice Analysis Toolbox \insertCite{Tsanas:2011cb,Tsanas:2012un,tsanas2013automatic}{superassp} applies a wide range of Matlab™-implemented voice
analysis procedures on a single sustained vowel (usually [a::]) and computes
339 acoustic quantities. This function calls a compiled application to
compute the outcome measures from each sustained vowel recording in a
directory, and collect the results into a \link[base:data.frame]{base::data.frame}.
}
\details{
The user should be aware that applying this procedure to a directory of sound files may take 9-40 times the total duration of the sound files to perform, depending on the machine and how the application was compiled.
The user should therefore make sure to capture the tibble once returned (or retrieve it immidiately from the \code{.Last.value} variable once the command completes).

Under the hood, the Voice Analysis Toolbox utilizes several other Matlab™ toolboxes which are also compiled into the runtime binary that performs the procedure.
The permissive open source licences of these published external toolboxes \insertCite{little2006nonlinear,little2007exploiting,brookes2011voicebox}{superassp} are gratefully acknowledged.
}
\references{
\insertAllCited{}
}
