% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_voice_analysis.R
\name{voice_analysis_info}
\alias{voice_analysis_info}
\title{Get voice_analysis system information}
\usage{
voice_analysis_info()
}
\value{
A list with system information and capabilities
}
\description{
Get voice_analysis system information
}
\examples{
\dontrun{
info <- voice_analysis_info()
print(info)
}
}
