% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_slicefunctions.R
\name{voice_analysis_toolkit}
\alias{voice_analysis_toolkit}
\title{Applies the Voice Analysis Toolbox on all vowel samples in a directory}
\usage{
voice_analysis_toolkit(
  listOfFiles,
  knownLossless = c("wav", "flac", "aiff", "aif")
)
}
\arguments{
\item{listOfFiles}{A list of input sound files. Currently, the function takes only .wav files.}

\item{knownLossless}{A vector of file extensions known to be associated with losslessly encoded speech data. Please note that this parameter is currently not used by the underlying code since only .wav files are processed.}
}
\value{
A named list of 339 voice measurements measurement values for the recording.
}
\description{
The Voice Analysis Toolbox \insertCite{Tsanas:2011cb,Tsanas:2012un,tsanas2013automatic}{superassp} applies a wide range of Matlab™-implemented voice
analysis procedures on a single sustained vowel (usually [a::]) and computes
339 acoustic quantities. This function calls a compiled application to
compute the outcome measures from each sustained vowel recording , and collect the results into a \link[base:list]{base::list}.
}
\details{
The user should be aware that applying this procedure to a directory of sound files may take 9-40 times the total duration of the sound files to perform, depending on the machine and how the application was compiled.
The user should therefore make sure to capture the tibble once returned (or retrieve it immidiately from the \code{.Last.value} variable once the command completes).

Under the hood, the Voice Analysis Toolbox utilizes several other Matlab™ toolboxes which are also compiled into the runtime binary that performs the procedure.
The permissive open source licences of these published external toolboxes \insertCite{little2006nonlinear,little2007exploiting,brookes2011voicebox}{superassp} are gratefully acknowledged.
}
\references{
\insertAllCited{}
}
