% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_voice_sauce.R
\name{voice_sauce_available}
\alias{voice_sauce_available}
\title{Check VoiceSauce Availability}
\usage{
voice_sauce_available()
}
\value{
Logical; TRUE if VoiceSauce is available, FALSE otherwise
}
\description{
Check if the VoiceSauce Python module is available.
}
\examples{
\\dontrun{
if (voice_sauce_available()) {
  vs <- lst_voice_sauce("audio.wav")
}
}

}
