% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice_sauce.R
\name{voice_sauce_info}
\alias{voice_sauce_info}
\title{VoiceSauce System Information}
\usage{
voice_sauce_info(print_info = TRUE)
}
\arguments{
\item{print_info}{Logical; if TRUE, print information to console}
}
\value{
Invisible list with system information
}
\description{
Display information about VoiceSauce installation and optimizations.
}
\examples{
\\dontrun{
voice_sauce_info()
}

}
