% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_voxit.R
\name{voxit_available}
\alias{voxit_available}
\title{Check if Voxit module is available}
\usage{
voxit_available()
}
\value{
Logical. TRUE if voxit module is available, FALSE otherwise
}
\description{
Check if Voxit module is available
}
\examples{
\dontrun{
if (voxit_available()) {
  features <- lst_voxit("audio.wav", alignmentFiles = "align.csv")
} else {
  install_voxit()
}
}

}
\seealso{
\code{\link{install_voxit}}, \code{\link{voxit_info}}
}
