% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat_internal_peak_slope.R
\name{wavelet_bands_wavelets_pkg}
\alias{wavelet_bands_wavelets_pkg}
\title{DWT band decomposition using 'wavelets' package}
\usage{
wavelet_bands_wavelets_pkg(s, n_levels)
}
\description{
DWT band decomposition using 'wavelets' package
}
\keyword{internal}
