% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat_internal_lpc.R
\name{zero_phase_hp_filt}
\alias{zero_phase_hp_filt}
\title{Zero-phase high-pass Butterworth filter}
\usage{
zero_phase_hp_filt(x, fs, f_p, f_s)
}
\arguments{
\item{x}{Input signal}

\item{fs}{Sampling frequency (Hz)}

\item{f_p}{Pass-band frequency (Hz)}

\item{f_s}{Stop-band frequency (Hz)}
}
\value{
Filtered signal
}
\description{
Zero-phase high-pass Butterworth filter
}
\keyword{internal}
