% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_ssff.R
\name{aperiodicities}
\alias{aperiodicities}
\title{Estimates aperiodicity of a speech signal}
\usage{
aperiodicities(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 70,
  maxF = 200,
  voiced_voiceless_threshold = 0.01,
  explicitExt = "wap",
  outputDirectory = NULL,
  toFile = TRUE
)
}
\arguments{
\item{listOfFiles}{A vector of file paths to wav files.}

\item{beginTime}{The start time of the section of the sound file that should be processed.}

\item{endTime}{The end time of the section of the sound file that should be processed.}

\item{windowShift}{The measurement interval (frame duration), in seconds.}

\item{minF}{Candidate f0 frequencies below this frequency will not be considered.}

\item{maxF}{Candidates above this frequency will be ignored.}

\item{explicitExt}{the file extension that should be used.}

\item{outputDirectory}{set an explicit directory for where the signal file will be written. If not defined, the file will be written to the same directory as the sound file.}

\item{toFile}{write the output to a file? The file will be written in  \code{outputDirectory}, if defined, or in the same directory as the soundfile.}
}
\value{
An SSFF track object containing two tracks (f0 and corr) that are
either returned (toFile == FALSE) or stored on disk.
}
\description{
Aperiodicity is estimated using function for assessing band-aperiodicities
using the Definitive Decomposition Derived Dirt-Cheap (D4C) algorithm
\insertCite{Morise.2016.10.1016/j.specom.2016.09.001}{superassp} implemented
in the WORLD vocoder
\insertCite{MORISE.2016.10.1587/transinf.2015edp7457}{superassp}. The the \link[=DIO]{dio}
\insertCite{morise2010rapid}{superassp} pitch algorithm is used to calculate
the periodic component.
}
\references{
\insertAllCited{}
}
