% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_core.R
\name{append_json_track_slice}
\alias{append_json_track_slice}
\title{Append a slice to JsonTrackObj}
\usage{
append_json_track_slice(obj, results, beginTime, endTime)
}
\arguments{
\item{obj}{JsonTrackObj}

\item{results}{New results to append}

\item{beginTime}{Start time}

\item{endTime}{End time}
}
\value{
Updated JsonTrackObj
}
\description{
Append a slice to JsonTrackObj
}
