% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_ssff.R
\name{dio}
\alias{dio}
\title{Compute f0 using the Harvest algorithm}
\usage{
dio(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 70,
  maxF = 200,
  voiced.voiceless.threshold = 0.01,
  explicitExt = "wd0",
  outputDirectory = NULL,
  toFile = TRUE
)
}
\arguments{
\item{listOfFiles}{A vector of file paths to wav files.}

\item{beginTime}{The start time of the section of the sound file that should be processed.}

\item{endTime}{The end time of the section of the sound file that should be processed.}

\item{windowShift}{The measurement interval (frame duration), in seconds.}

\item{minF}{Candidate f0 frequencies below this frequency will not be considered.}

\item{maxF}{Candidates above this frequency will be ignored.}

\item{voiced.voiceless.threshold}{Threshold for voiced/unvoiced decision. Can
be any value >= 0, but 0.02 to 0.2 is a reasonable range. Lower values will
cause more frames to be considered unvoiced (in the extreme case of
\code{threshold=0}, almost all frames will be unvoiced).}

\item{explicitExt}{the file extension that should be used.}

\item{outputDirectory}{set an explicit directory for where the signal file will be written. If not defined, the file will be written to the same directory as the sound file.}

\item{toFile}{write the output to a file? The file will be written in  \code{outputDirectory}, if defined, or in the same directory as the soundfile.}
}
\value{
An SSFF track object containing two tracks (f0 and corr) that are
either returned (toFile == FALSE) or stored on disk.
}
\description{
The DIO algorithm \insertCite{morise2010rapid}{superassp} was developed for
the WORLD vocoder
\insertCite{MORISE.2016.10.1587/transinf.2015edp7457}{superassp} and aims
provide a fast estimate of the f0 contour.
}
\references{
\insertAllCited{}
}
