% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdata_extensions.R
\name{get_definedtracks}
\alias{get_definedtracks}
\title{Get available tracks for a function or an SSFF file}
\usage{
get_definedtracks(x)
}
\arguments{
\item{x}{The name of function defined to output trackdata, or the full path to a file that can be read using \link[wrassp:read.AsspDataObj]{wrassp::read.AsspDataObj}.}
}
\value{
A vector of tracks that the function is defined to return, or are contained within the file.
}
\description{
Get available tracks for a function or an SSFF file
}
\examples{
get_definedtracks("forest")
get_definedtracks("praat_formant_burg")

}
