% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff.R
\name{harmonics}
\alias{harmonics}
\title{Compute the harmonic structure from f0 measurements}
\usage{
harmonics(f0, n = 10, explicitExt = "har", toFile = TRUE)
}
\arguments{
\item{f0}{An f0 track, either as an SSFF object or as the name of an SSFF formatted file.}

\item{n}{The number of harmonics to compute.}

\item{explicitExt}{The output file extension.}

\item{toFile}{boolean;Should the SSFF track be returned or stored on disc?}
}
\value{
The SSFF track object, if required. \code{NULL} otherwise.
}
\description{
This function takes a pre-computed f0 track and derive \code{n} harmonic tracks
from it so that the vector of f0 values are now a matrix with \code{n} columns.
Each column then encode the \code{n}th harmonic values.
}
