% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrassp_AsspTypes.R
\name{knownLossless}
\alias{knownLossless}
\title{Procedure reporting lossless file formats}
\usage{
knownLossless()
}
\description{
This procedure just reports back file extensions that are known to
predominately contain losslessly encoded sound data.
}
\author{
Fredrik Nylén
}
