% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrassp_rfcana.R
\name{larana}
\alias{larana}
\title{Linear Prediction analysis using autocorrelation the Durbin recursion and log area ratio output.}
\usage{
larana(listOfFiles = NULL, 
  beginTime = 0.0, 
  centerTime = FALSE, 
  endTime = 0.0, 
  windowShift = 5.0, 
  windowSize = 20.0, 
  effectiveLength = TRUE, 
  window = 'BLACKMAN', 
  order = 0, 
  preemphasis = -0.95,
  toFile = TRUE,
  explicitExt=NULL,
  outputDirectory = NULL, 
  assertLossless = NULL, 
  logToFile = FALSE,
  keepConverted=FALSE,
  verbose = TRUE)
}
\arguments{
\item{listOfFiles}{vector of file paths to be processed by function}

\item{beginTime}{the time point (in seconds) of the start of the analysed
interval. A NULL or 0 is interpreted as the start of the signal file.
If a vector of time points is supplied, the length of that vector needs
to correspond with the length of \code{listOfFiles}.}

\item{centerTime}{sets a single-frame analysis time point (in seconds).
Overrides \code{beginTime}, \code{endTime} and \code{windowShift} parameters.}

\item{endTime}{the time point (in seconds) of the end of the analysed
interval. A NULL or 0 is interpreted as the end of the signal file.
If a vector of time points is supplied, the length of that vector needs
to correspond with the length of \code{listOfFiles}.}

\item{windowShift}{the amount of time (in ms) that the analysis window will
be shifted between analysis frames}

\item{windowSize}{the analysis window size (in ms); overrides the effect of
the \code{effectiveLength} parameter}

\item{effectiveLength}{make window size effective rather than exact}

\item{window}{= the analysis window function type ("BLACKMAN" by default).
See \link{AsspWindowTypes} for a list of supported window types.}

\item{analysisOrder}{the analysis order. The \code{NULL} or \code{0} sets the analysis
order to the sample rate (in kHz) + 3, so that a signal with a 16000 Hz
sampling rate will be analysed using an \code{analysisOrder} of 19.}

\item{preemphasis}{= \if{html}{\out{<val>}}: set pre-emphasis factor to \if{html}{\out{<val>}} (default: -0.95)}

\item{toFile}{Should the function write the results to a file, with the
(default) file extension (\code{TRUE}) or returned as a list of
\link{AsspDataObj} objects (\code{FALSE})?}

\item{explicitExt}{the file extension will be used when
result files are written (\code{toFile=TRUE}), but the file extension can be
set to something else using this function argument.}

\item{outputDirectory}{directory in which output files are stored. Defaults
to NULL which means that the result file will be stored in the same
directory as the input file.}

\item{verbose}{display verbose information about processing steps taken, as
well as progress bars.}

\item{assertLossless}{an optional list of file extensions that the user wants to assert
contains losslessly encoded signals data.}

\item{logToFile}{whether to log commands to a separate logfile in the
\code{outputDirectory}. Logging will otherwise be in the function-specific logging
namespace of \link{logger} and will be put wherever this namespace is defined to place its output.
See \link[logger:log_appender]{logger::log_appender} for details.}
}
\value{
The number of successfully written files (if \code{toFile=TRUE}), or a vector of \code{AsspDataObj} objects (if \code{toFile=FALSE}).
}
\description{
Linear Prediction analysis of \code{listOfFiles} using the
autocorrelation method and the Durbin recursion.
This program calculates the RMS amplitudes of the input
and residual signal in dB and, and log area ratios using algorithms implmented in \emph{libassp} \insertCite{s5h}{superassp}.
Input signals not in a file format natively
supported will be converted before the autocorrelation functions are
computed. The conversion process will display warnings about input files
that are not in known losslessly encoded formats.
}
\details{
The results will be will be written to an SSFF formated file with the base
name of the input file and extension \code{.lar} with tracks \code{RMS[dB]},\code{gain[dB]},and \code{LAR}.

The function is a re-write of the \link[wrassp:rfcana]{wrassp::rfcana} function with log area ratio output, but
with media pre-conversion, better checking of preconditions such as the
input file existance, structured logging, and the use of a more modern
framework for user feedback.

The native file type of this function is "wav" files (in "pcm_s16le"
format), SUNs "au", NIST, or CSL formats (kay or NSP extension). Input
signal conversion, when needed, is done by
\href{https://ffmpeg.org/libavcodec.html}{libavcodec} and the excellent \link[av:encoding]{av::av_audio_convert}
wrapper function
}
\note{
This function is not considered computationally expensive enough to require caching of
results if applied to many signals. However, if the number of signals it will be applied to
is \emph{very} long, then caching of results may be warranted.
}
\examples{
# get path to audio file
path2wav <- list.files(system.file("samples","sustained", package = "superassp"), pattern = glob2rx("a1.wav"), full.names = TRUE)

# perform linear prediction analysis
res <- rfcana(path2wav, toFile=FALSE)

# plot reflection coefficients
matplot(seq(0,numRecs.AsspDataObj(res) - 1) / rate.AsspDataObj(res) + 
          attr(res, 'startTime'), 
        res$LAR, 
        type='l', 
        xlab='time (s)', 
        ylab='Log area ratios')
        
}
\references{
\insertAllCited{}
}
\seealso{
\link[wrassp:rfcana]{wrassp::rfcana}

\link{AsspWindowTypes}

\link[av:encoding]{av::av_audio_convert}
}
\author{
Raphael Winkelmann

Lasse Bombien

Fredrik Nylén
}
