% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_python_opensmile_emobase.R
\name{lst_emobase}
\alias{lst_emobase}
\title{Compute the emobase openSMILE feature set}
\usage{
lst_emobase(listOfFiles, ...)
}
\arguments{
\item{listOfFiles}{The full path to the sound file.}

\item{beginTime}{The starting time of the section of the sound files that
should be analysed.}

\item{endTime}{The end time of the section of the sound files that should be
analysed.}

\item{explicitExt}{The file extension of the slice file where the results
should be stored.}
}
\value{
A list of 988 acoustic values, with the names as reported by
openSMILE.
}
\description{
This function applies the emobase openSMILE \insertCite{Eyben:2010fq}{superassp} configuration to compute 988
acoustic features reasoned to be part of perception of emotion.
}
\references{
\insertAllCited{}
}
