% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat_slicefunctions.R
\name{praat_dsi}
\alias{praat_dsi}
\title{Compute the Dysphonia Severity Index}
\usage{
praat_dsi(
  softDF,
  highpitchDF,
  maxprolongedDF,
  stableDF = NULL,
  use.calibration = FALSE,
  db.calibration = 10,
  speaker.name = NULL,
  speaker.ID = NULL,
  speaker.dob = NULL,
  session.datetime = NULL,
  pdf.path = NULL,
  overwrite.pdfs = FALSE,
  praat_path = NULL
)
}
\arguments{
\item{softDF}{A data.frame containing the file path and start and end times
of samples indicating the softest voice production of the speaker.}

\item{highpitchDF}{A data.frame containing the file path and start and end
times of samples from which the highest f0 of the speaker can be computed.}

\item{maxprolongedDF}{A data.frame containing the file path and start and end
times of samples indicating maximally prolonged vowels of the speaker.}

\item{stableDF}{An optional data.frame containing the file path and start and
end times of samples indicating a stable sustained vowel. If not provided,
the \code{maxprolongedDF} will be used instead.}

\item{use.calibration}{Should a calibration factor be added to measured
intensities before computing DSI?}

\item{db.calibration}{The number of dB:s to be added to measured intensity
values before computing DSI.}

\item{speaker.name}{An optional name of the speaker. Only used when PDF
output is produced.}

\item{speaker.ID}{An ID of the speaker. This will be included in the list
output.}

\item{speaker.dob}{An optional date of birth of the speaker. Only used when
PDF output is produced.}

\item{session.datetime}{A string indicating the session date (and time). Only
used when PDF output is produced.}

\item{pdf.path}{A path where a PDF report file will be produced.}

\item{overwrite.pdfs}{Should existing PDF files be overwritten in the PDF
output directory? Defaults to a safe behavior where older PDFs are not
overwritten.}

\item{praat_path}{An explicit path to the Praat binary.}
}
\value{
A list with the following fields:
\describe{
\item{ID}{The speaker speaker + session identifier of the output}
\item{Maximum.phonation.time}{The speakers maximum phonation time.}
\item{Softest.intensity.of.voiced.speech}{The intensity of the participants
softest voice (in dB)}
\item{Maximum.fundamental.frequency}{The maximum f0
frequency (in Hz)}
\item{Jitter.ppq5}{The five-point Period Pertubation
Quotient measurement (in \%)}
\item{Dysphonia.Severity.Index}{The resulting
Dysphonia Severity Index }
}

A meta-analysis in which 1330 healthy participants (aged 17.3−94) were included indicated an average DSI of 3.05 in healthy speakers, with a conﬁdence level  between
2.13 and 3.98 \insertCite{Sobol.2020.10.1016/j.jvoice.2020.04.010}{superassp}.
}
\description{
This function computes the Dysphonia Severity Index (DSI)
\insertCite{Wuyts:2000vb}{superassp} as implemented into Praat by
\insertCite{Maryn.2017.10.1016/j.jvoice.2017.01.002;textual}{superassp}. The
user is asked to supply data.frames indicating the "absolute_file_path" as
well as "start", and "end" times for parts of sound files that together form
the basis for the DSI computation for a speaker.
}
\details{
The user has to indicate at least one sample in which the participant speaks
as softly as possible (\code{softDF}), at least one sample from which the maximum
f$_0$ could be deduced (\code{highpitchDF}), at least one sample where a single
vowel is maximally prolonged (\code{maxprolongedDF}). The user can also provide a
sample of a maximally stable vowel (\code{stableDF}), but if such a sample is not
provided then the (\code{maxprolongedDF}) sample will be reused instead for the
computation of the DSI sub-component Jitter. The user may submit multiple
sound samples for all these sets of acoustic inputs, and the sounds files
will then be combined before the DSI sub components are computed. The largest
Maximum performance time will be used.
}
\references{
\insertAllCited{}
}
