% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat.R
\name{praat_formant_burg}
\alias{praat_formant_burg}
\title{Use Praat to compute a formant track using the burg method.}
\usage{
praat_formant_burg(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 0,
  numFormants = 5,
  maxhzformant = 5500,
  windowSize = 0.025,
  preemphasis = 50,
  window = "hanning",
  relativeWidth = 1,
  toFile = TRUE,
  explicitExt = "fms",
  outputDirectory = NULL,
  praat_path = NULL
)
}
\arguments{
\item{listOfFiles}{a vector of wav file paths to be processed by function.}

\item{beginTime}{the time where processing should end (in s) The default is 0 (zero) which means that the computation of formants will start at the start of the sound file.}

\item{endTime}{the time where processing should end (in s) The default is 0 (zero) which means that formants will be computed up to the end of the file.}

\item{windowShift}{the analysis window shift length (in ms).}

\item{numFormants}{the number of formants that the analysis should try to find}

\item{maxhzformant}{praat will try to find formants only up to this frequency in the spectrum.}

\item{windowSize}{the analysis window length (in ms).}

\item{preemphasis}{the frequency from which a preemphasis will be applied..}

\item{window}{the analysis window function used when extracting part of a sound file for analysis. De faults to "Hanning".}

\item{relativeWidth}{the relative width of the windowing function used.}

\item{toFile}{write the output to a file? The file will be written in  `outputDirectory`, if defined, or in the same directory as the soundfile.}

\item{explicitExt}{the file extension that should be used.}

\item{outputDirectory}{set an explicit directory for where the signal file will be written. If not defined, the file will be written to the same directory as the sound file.}

\item{praat_path}{give an explicit path for Praat. If the praat}
}
\value{
a list of
}
\description{
This function asks Praat to compute a formant track using the burg method, and the result is converted in an SSFF file.
This function should be a drop-in replacement for the \code{\link[wrassp]{forest}} function, but with some additional
arguments. If the function cannot find the Praat binary automatically, you have to give an explicit path (e.g. "/Applications/Praat.app/Contents/MacOS/Praat" if you placed Praat in the Applications folder on your Mac). 
You can check whether you need to supply a explicit path using the \code{\link{get_praat}} or 
\code{\link{has_praat}} functions.
}
