% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat_ssff.R
\name{praat_pitch}
\alias{praat_pitch}
\title{Compute f0 tracks using Praat}
\usage{
praat_pitch(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 50,
  maxF = 300,
  max.f0.candidates = 15,
  very.accurate = TRUE,
  silence.threshold = 0.03,
  voicing.threshold = 0.45,
  octave.cost = 0.01,
  octave.jump.cost = 0.35,
  voiced.voiceless.cost = 0.14,
  windowShape = "Gaussian1",
  relativeWidth = 1,
  toFile = TRUE,
  explicitExt = "pf0",
  outputDirectory = NULL,
  verbose = FALSE,
  praat_path = NULL
)
}
\arguments{
\item{listOfFiles}{}

\item{beginTime}{}

\item{endTime}{}

\item{windowShift}{The measurement interval (frame duration), in seconds. If you supply 0, Praat will use a time step of 0.75 / (pitch floor), e.g. 0.01 seconds if the pitch floor is 75 Hz; in this example, Praat computes 100 pitch values per second.}

\item{minF}{candidates below this frequency will not be recruited. This parameter determines the effective length of the analysis window: it will be 3 longest periods long, i.e., if the pitch floor is 75 Hz, the window will be effectively 3/75 = 0.04 seconds long. Note that if you set the time step to zero, the analysis windows for consecutive measurements will overlap appreciably: Praat will always compute 4 pitch values within one window length, i.e., the degree of oversampling is 4.}

\item{maxF}{Candidates above this frequency will be ignored.}

\item{max.f0.candidates}{The maximum numbef of f0 candidates to consider}

\item{very.accurate}{If FALSE, the window is a Hanning window with a physical length of 3 / (pitch floor). If TRUE, the window is a Gaussian window with a physical length of 6 / (pitch floor), i.e. twice the effective length.}

\item{silence.threshold}{Frames that do not contain amplitudes above this threshold (relative to the global maximum amplitude), are probably silent.}

\item{voicing.threshold}{The strength of the unvoiced candidate, relative to the maximum possible autocorrelation. To increase the number of unvoiced decisions, increase this value.}

\item{octave.cost}{The degree of favoring of high-frequency candidates, relative to the maximum possible autocorrelation. This is necessary because even (or: especially) in the case of a perfectly periodic signal, all undertones of f0 are equally strong candidates as f0 itself. To more strongly favour recruitment of high-frequency candidates, increase this value.}

\item{octave.jump.cost}{Degree of disfavoring of pitch changes, relative to the maximum possible autocorrelation. To decrease the number of large frequency jumps, increase this value. In contrast with what is described by \insertCite{Boersma1993}{superassp}, this value will be corrected for the time step: multiply by 10ms / windowShift to get the value in the way it is used in the formulas in the article.}

\item{voiced.voiceless.cost}{Degree of disfavoring of voiced/unvoiced transitions, relative to the maximum possible autocorrelation. To decrease the number of voiced/unvoiced transitions, increase this value. In contrast with what is described in the article, this value will be corrected for the time step: multiply by 10 ms / windowShift to get the value in the way it is used in the formulas in \insertCite{Boersma1993}{superassp}.}

\item{windowShape}{the analysis window function used when extracting part of a sound file for analysis. De faults to "Hanning".}

\item{relativeWidth}{the relative width of the windowing function used.}

\item{toFile}{write the output to a file? The file will be written in  \code{outputDirectory}, if defined, or in the same directory as the soundfile.}

\item{explicitExt}{the file extension that should be used.}

\item{outputDirectory}{set an explicit directory for where the signal file will be written. If not defined, the file will be written to the same directory as the sound file.}

\item{verbose}{Not implemented. Only included here for compatibility.}

\item{praat_path}{give an explicit path for Praat. If the praat}
}
\value{

}
\description{
This function calls Praat to compute f0 tracks. Both the auto-correlation and cross-correlation methods are used, and the results are stored in separate fields in the returned SSFF track object.
Most arguments to the function map directly to formal arguments to the underlying Praat procedure, and the description of these are therefore replicated here. See the Praat manual for more information.
}
\references{
\insertAllCited{}
}
