% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat.R
\name{praat_sauce}
\alias{praat_sauce}
\title{Call the 'praat_sauce' analysis bundle to generate SSFF tracks}
\usage{
praat_sauce(
  listOfFiles,
  beginTime = NULL,
  endTime = NULL,
  windowSize = 0.025,
  minF = 50,
  maxF = 300,
  timeStep = 0.005,
  formantTracking = TRUE,
  numFormants = 5,
  maxFormantHz = 5000,
  nominalF1 = 500,
  nominalF2 = 1500,
  nominalF3 = 2500,
  preEmphFrom = 50,
  useBandwidthFormula = FALSE,
  channel = 1,
  resample_to_16k = TRUE,
  toFile = TRUE,
  explicitExt = "psa",
  outputDirectory = NULL,
  verbose = FALSE,
  praat_path = NULL
)
}
\arguments{
\item{listOfFiles}{A vector of file names}

\item{beginTime}{The time in the sound file where analysis should start.}

\item{endTime}{The last time point to be included in the analysed sample. If zero (0), the sound file will be included until the end.}

\item{windowSize}{The analysis window length.}

\item{minF}{The minimal f0 to search for.}

\item{maxF}{The maximum f0 to search for.}

\item{timeStep}{Time step between analysis windows. Defaults to every 5ms (0.005s).}

\item{formantTracking}{Boolean; Should the formant tracking abilities of Praat be used? Defaults to TRUE. If disabled, the raw window-by-window formant values will be used.}

\item{numFormants}{The number of formants to be found within the frequency space.}

\item{maxFormantHz}{The cutoff frequency used when finding the `numFormants` formants.}

\item{nominalF1}{The nominal F1 used in formant tracking.}

\item{nominalF2}{The nominal F2 used in formant tracking.}

\item{nominalF3}{The nominal F3 used in formant tracking.}

\item{preEmphFrom}{The frequency from which pre-emphasis will be applied.}

\item{useBandwidthFormula}{Should the bandwidth calculation metod of \insertCite{Hawks.1995.10.1121/1.412986;textual}{superassp} be used, instead of Praat's internal algorithm. Defaults to TRUE (use Hawks & Miller's method).}

\item{channel}{Which channel to use analyse. Defaults to the first channel.}

\item{resample_to_16k}{Resample the signal to 16000 Hz before processing? Defaults to TRUE.}

\item{toFile}{Should the SSFF track file be written to disk? Defaults to true, which by default means that it will be written as a file with a `explicitExt` extension  next to the sound file.}

\item{explicitExt}{The default file extension to use for the SSFF file.}

\item{outputDirectory}{A path to an alternative output directory.}

\item{verbose}{Verbose output. Not currently used.}

\item{praat_path=NULL}{}
}
\value{
This function builds an SSFF track object and writes it to disk, or returns it (`toFile==FALSE`). The track object will
contain tracks with these fields:
\describe{
  \item{f0}{A track of the fundamental frequency computed by Praat from the signal as part of the PraatSauce analysis.}
  \item{fm}{Computed formant frequency measures, optionally tracked, one column per formant. }
  \item{bw}{Computed formant bandwiths, one column per formant.}
  \item{H}{The amplitudes of the first three harmonics, computed without accounting for the influence of adjecent formants (uncorrected).}
  \item{Hc}{The amplitudes of the first three harmonics, in which the influence of adjecent formants has been accounted for (corrected).}
  \item{A}{The amplitudes of the three harmonics that are closest to the center frequency of the first three formants, computed without accounting for the influence of the adjecent formants (uncorrected).}
  \item{Ac}{The amplitudes of the three harmonics that are closest to the center frequency of the first three formants, in which the influence of the adjecent formants has been accounted for (corrected).}
  \item{HH}{The  differences between the (uncorrected) amplitudes of the first and second (column 1) and second and fourth (column 2).}
  \item{HHc}{The  differences between the amplitudes of the first and second (column 1) and second and fourth (column 2), computed with correction with regards to neighbouring formants.}
  \item{HA}{The  differences between the (uncorrected) amplitudes of the harmonics closest to the center frequencies of the first and second (column 1) and second and fourth (column 2) formants.}
  \item{HAc}{The  differences between the amplitudes of the harmonics closest to the center frequencies of the first and second (column 1) and second and fourth (column 2) formants, corrected for the influence of the adjecent formants themselves.}
  \item{cpp}{A track of containing the smoothed Cepstral Peak Prominence \insertCite{Fraile.2014.10.1016/j.bspc.2014.07.001,Hillenbrand.1994.10.1044/jshr.3704.769}{superassp} across the acoustic signal.}
  \item{hnr}{Harmonic-to-noise ratios as defined by \insertCite{Krom.1993.10.1044/jshr.3602.254}{superassp} for frequencies 0-500 Hz, 0-1500 Hz, 0-2500 Hz, and 0-3500 Hz, respectively. (four columns)}
}
}
\description{
This function applies the \code{praat_sauce} bundle of Praat scripts
on a single file and puts the resulting data as tracks in an SSFF file.
By default, all analyses are applied to a windowed portion of the signal
every 5ms.
}
\examples{
\dontrun{

}
}
\references{
\insertAllCited{}
}
