% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praat_slicefunctions.R
\name{praat_voice_tremor}
\alias{praat_voice_tremor}
\title{Computes 18 measures of (vocal) tremor}
\usage{
praat_voice_tremor(
  listOfFiles,
  beginTime = NULL,
  endTime = NULL,
  selectionOffset = NULL,
  selectionLength = 3,
  windowShape = "Gaussian1",
  relativeWidth = 1,
  minF = 75,
  maxF = 600,
  windowShift = 5,
  max_period_factor = 1.3,
  max_ampl_factor = 1.6,
  silence_threshold = 0.03,
  voicing_threshold = 0.45,
  octave_cost = 0.01,
  octave_jump_cost = 0.35,
  voiced_unvoiced_cost = 0.14,
  envelope.amplitude = TRUE,
  min.tremor.hz = 1.5,
  max.tremor.hz = 15,
  contour.magnitude.threshold = 0.01,
  tremor.cyclicality.threshold = 0.15,
  freq.tremor.octave.cost = 0.01,
  ampl.tremor.octave.cost = 0.01,
  na.zero = FALSE,
  praat_path = NULL
)
}
\arguments{
\item{listOfFiles}{The full path of the sound file.}

\item{beginTime}{The time point (in s) in the sound file where the sustained vowel starts. If \code{NULL}, the start of the sound file will also be viewed as the start of the sustained vowel production.}

\item{endTime}{The time point (in s) in the sound file where the sustained vowel ends. If \code{NULL}, everyting up until the end of the sound file will be considered part of the sustained vowel.}

\item{selectionOffset}{An optional offset to be added to the time of the sustained vowel production when determining where the start of the extracted portion of the vowel.}

\item{selectionLength}{An optional (maximal) length of the selection.}

\item{windowShape}{The window shape used for extracting the vowel. May be one of "rectangular", "triangular", "parabolic", "Hanning", "Hamming", "Gaussian1", "Gaussian2", "Gaussian3", "Gaussian4", "Gaussian5", "Kaiser1", and "Kaiser2".}

\item{relativeWidth}{The relative width of the window used for extracting the vowel portion.}

\item{minF}{The minimum pitch (f~0~) to be considered.}

\item{maxF}{The maximum pitch (f~0~) to be considered.}

\item{windowShift}{The number of milliseconds between analysis window centers.}

\item{max_period_factor}{The larges possible differences between consecutive intervals that will be used in computing jitter. Please consult the Praat manual for further information.}

\item{max_ampl_factor}{The larges possible differences between consecutive intervals that will be used in computing schimmer Please consult the Praat manual for further information.}

\item{silence_threshold}{The silence threshold. Please consult the Praat manual for further information.}

\item{voicing_threshold}{The voicing  threshold. Please consult the Praat manual for further information.}

\item{octave_cost}{The octave cost. Please consult the Praat manual for further information.}

\item{octave_jump_cost}{The octave jump cost. Please consult the Praat manual for further information.}

\item{voiced_unvoiced_cost}{The cost for voiced to unvoiced change detection. Please consult the Praat manual for further information.}

\item{na.zero}{Should measurements of voice tremor that could not be determined be returned as a zero value, or as an \code{NA} (the default).}

\item{praat_path}{An optional explicit path to the Praat binary. Not usually required.}
}
\value{
A list of voice tremor measurements :
\describe{
\item{FCoM}{frequency contour magnitude}
\item{FTrC}{(maximum) frequency tremor cyclicality}
\item{FMon}{number of frequency modulations above thresholds}
\item{FTrF \link{Hz}}{(strongest) frequency tremor frequency}
\item{FTrI \link{\%}}{frequency tremor intensity index}
\item{FTrP}{frequency tremor power index}
\item{FTrCIP}{frequency tremor cyclicality intensity product}
\item{FTrPS}{frequency tremor product sum}
\item{FCoHNR\link{dB}}{frequency contour harmonicity-to-noise ratio}
\item{ACoM}{amplitude contour magnitude}
\item{ATrC}{(maximum) amplitude tremor cyclicality}
\item{AMoN}{number of amplitude modulations above thresholds}
\item{ATrF \link{Hz}}{(strongest) amplitude tremor frequency}
\item{ATrI \link{\%}}{amplitude tremor intensity index}
\item{ATrP}{amplitude tremor power index}
\item{ATrCIP}{amplitude tremor cyclicality intensity product }
\item{ATrPS}{amplitude tremor product sum}
\item{ACoHNR\link{dB}}{amplitude contour harmonicity-to-noise ratio}
}
}
\description{
This function calls Praat and applies the "tremor 3.05" package
\insertCite{tremor305}{superassp} to compute measures of
voice tremor \insertCite{Bruckl2017}{superassp} on a
sustained vowel. The user may compute the measures for the entire
sample or identify a portion of the sound file where the sustained is
produced (\code{beginTime} and \code{endTime} arguments). The user may additionally
specify an offset (\code{selectionOffset}) from the \code{beginTime} where measurements
should start, as well as a (maximum) selection length. Using these arguments
the user may mark where all sustained vowels are produced in a database, and
apply this function across only a 2 second sample  from 1 second into the
vowel, and thus avoid vowel initial effects in phonation affecting measurements.
}
\references{
\insertAllCited{}
}
