% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrassp_fileHelper.R
\name{prepareFiles}
\alias{prepareFiles}
\title{Prepare a file path for signal processing functions}
\usage{
prepareFiles(listOfFiles)
}
\arguments{
\item{listOfFiles}{The list of file names to process}
}
\value{
A normalised list of filenames
}
\description{
Normalise a list of filenames so that they can be passed to a signal processing function
}
\author{
Raphael Winkelmann
}
