% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_ssff_optimized.R
\name{process_swipe_single}
\alias{process_swipe_single}
\title{Internal function to process single file with SWIPE}
\usage{
process_swipe_single(
  soundFile,
  beginTime,
  endTime,
  windowShift,
  minF,
  maxF,
  voicing.threshold,
  explicitExt,
  outputDirectory,
  toFile
)
}
\description{
Internal function to process single file with SWIPE
}
\keyword{internal}
