% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_ssff.R
\name{reaper}
\alias{reaper}
\title{Extract f0 tracks using the REAPER algorithm (Optimized)}
\usage{
reaper(
  listOfFiles = NULL,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 40,
  maxF = 500,
  unvoiced_cost = 0.9,
  high.pass = TRUE,
  hilbert.transform = FALSE,
  explicitExt = "rp0",
  outputDirectory = NULL,
  toFile = TRUE,
  assertLossless = NULL,
  logToFile = FALSE,
  keepConverted = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{listOfFiles}{A vector of file paths to wav files.}

\item{beginTime}{The start time of the section of the sound file that should be processed.}

\item{endTime}{The end time of the section of the sound file that should be processed.}

\item{windowShift}{The measurement interval (frame duration), in seconds.}

\item{minF}{Minimum f0 in Hz}

\item{maxF}{Maximum f0 in Hz}

\item{unvoiced_cost}{Cost for unvoiced segments (0-1, higher = more f0)}

\item{high.pass}{Perform high-pass filtering?}

\item{hilbert.transform}{Use Hilbert transform for phase distortion?}

\item{explicitExt}{File extension for output files}

\item{outputDirectory}{Output directory path (NULL for same as input)}

\item{toFile}{Write output to file (TRUE) or return object (FALSE)}

\item{assertLossless}{an optional list of file extensions that the user wants to assert
contains losslessly encoded signals data.}

\item{logToFile}{whether to log commands to a separate logfile in the
\code{outputDirectory}. Logging will otherwise be in the function-specific logging
namespace of \link[logger:logger]{logger::logger} and will be put wherever this namespace is defined to place its output.
See \link[logger:log_appender]{logger::log_appender} for details.}

\item{keepConverted}{whether to keep converted files}

\item{verbose}{display verbose information about processing steps taken, as
well as progress bars.}
}
\value{
If \code{toFile} is \code{FALSE}, returns list of AsspDataObj objects.
If \code{toFile} is \code{TRUE}, returns count of successfully processed files.
}
\description{
Robust Epoch And Pitch EstimatoR (REAPER) algorithm with optimized
file processing following the forest function template.
}
