% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_ssff.R
\name{reaper_pm}
\alias{reaper_pm}
\title{Extract pitch marks using the REAPER algoritm}
\usage{
reaper_pm(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 10,
  minF = 40,
  maxF = 500,
  unvoiced_cost = 0.9,
  high.pass = TRUE,
  hilbert.transform = FALSE,
  explicitExt = "rpm",
  outputDirectory = NULL,
  toFile = TRUE,
  conda.env = NULL
)
}
\arguments{
\item{listOfFiles}{A vector of file paths to wav files.}

\item{beginTime}{The start time of the section of the sound file that should be processed.}

\item{endTime}{The end time of the section of the sound file that should be processed.}

\item{windowShift}{The measurement interval (frame duration), in seconds.}

\item{minF}{Minimum f0 in Hz}

\item{maxF}{Maximum f0 in Hz}

\item{unvoiced_cost}{Cost for unvoiced segments (0-1, higher = more f0)}

\item{high.pass}{Perform high-pass filtering?}

\item{hilbert.transform}{Use Hilbert transform for phase distortion?}

\item{explicitExt}{File extension for output files}

\item{outputDirectory}{Output directory path (NULL for same as input)}

\item{toFile}{Write output to file (TRUE) or return object (FALSE)}
}
\value{
An SSFF track object containing two tracks (f0 and corr) that are
either returned (toFile == FALSE) or stored on disk.
}
\description{
Robust Epoch And Pitch EstimatoR (REAPER) algorithm
\insertCite{talkin2019reaper}{superassp} uses an EpochTracker class to
simultaneously estimate the location of voiced-speech "epochs" or glottal
closure instants (GCI), voicing state (voiced or unvoiced) and fundamental
frequency (F0 or "pitch"). This function returns the voicing state of each
windowed \code{windowShift} (ms) portion of the signal.
}
\details{
DC bias and low-frequency noise are removed by high-pass filtering, and the
signal is converted to floating point. If the input is known to have phase
distortion that is impacting tracker performance, a Hilbert transform,
optionally done at this point, may improve performance.

The function uses the python library \code{pyreaper} combined with the R
package \link[reticulate:reticulate]{reticulate::reticulate} to compute the tracks, and the user therefore has to
make sure that \code{pyreaper} and python is available on the machine. It is
recommended to set up an anaconda ("conda") environment for the superassp
library, like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{conda create conda create --prefix -n pysuperassp python=3.8 
conda activate pysuperassp 
pip install librosa
pip install pyreaper 
#Not used by
this function but by other functions in this package 
pip install pysptk 
}\if{html}{\out{</div>}}
}
\references{
\insertAllCited{}
}
