% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superassp-package.R
\docType{package}
\name{superassp-package}
\alias{superassp-package}
\title{superassp: speech signal processing using various framworks using a wrassp-like interface}
\description{
This package bundles together routines that utilizes Praat, matlab or python
to do the signal processing. The output of the original routines are then
wrapped into a common interface that outputs an SSFF file or returns a wrassp compatible AsspDataObj object.
}
