% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_ssff.R
\name{torch_pitch}
\alias{torch_pitch}
\title{Pitch tracking using the torch pitch tracker}
\usage{
torch_pitch(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 10,
  windowSize = 30,
  minF = 70,
  maxF = 200,
  explicitExt = "tpi",
  outputDirectory = NULL,
  toFile = TRUE
)
}
\arguments{
\item{listOfFiles}{Vector of file paths to process}

\item{beginTime}{Start time in seconds (default: 0.0)}

\item{endTime}{End time in seconds (default: 0.0 = end of file)}

\item{windowShift}{Frame shift in milliseconds (default: 10.0)}

\item{minF}{Minimum F0 in Hz (default: 60.0)}

\item{maxF}{Maximum F0 in Hz (default: 400.0)}

\item{explicitExt}{Output file extension (default: "f0")}

\item{outputDirectory}{Output directory (default: NULL = same as input)}

\item{toFile}{Write results to file (default: TRUE)}
}
\value{
An SSFF track object containing two tracks (f0 and pitch) that are either returned (toFile == FALSE) or stored on disk.
}
\description{
This function estimates pitch by normalized cross-correlation function (NCCF) and
median smoothing, as implemented in the torchaudio
\insertCite{yang2021torchaudio}{superassp} library. The exact algorithm is undisclosed by the implementing library
but approach likely builds on earlier implementations that use NCCFs \insertCite{talkin1995robust,Kasi.2002.10.1109/icassp.2002.5743729}{superassp}
including the \link[superassp::rapt]{RAPT} algorithm.
}
\references{
\insertAllCited{}
}
\seealso{
rapt
}
