% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_c_assp_cepstrum.R
\name{trk_cepstrum}
\alias{trk_cepstrum}
\title{Performs a short-term cepstral analysis of the signal (Rcpp-optimized)}
\usage{
trk_cepstrum(listOfFiles, ...)
}
\arguments{
\item{resolution}{= \if{html}{\out{<freq>}}: set FFT length to the smallest value which
results in a frequency resolution of \if{html}{\out{<freq>}} Hz or better (default: 40.0)}

\item{fftLength}{= \if{html}{\out{<num>}}: set FFT length to \if{html}{\out{<num>}} points (overrules default
and 'resolution' option)}
}
\value{
The number of successfully written files (if \code{toFile=TRUE}), or a vector of \code{AsspDataObj} objects (if \code{toFile=FALSE}).
}
\description{
This function performs a cepstral analysis \insertCite{Oppenheim.2004.10.1109/msp.2004.1328092,Childers.1977.10.1109/proc.1977.10747}{superassp} on the signals in  \code{listOfFiles}
using the Fast Fourier Transform. The number of
coefficients per output record will also equal the
FFT length / 2 + 1 (which means that it will not be mirrored).

The results will be will be written to an SSFF formated file with the base
name of the input file and extension \emph{.cep} in a track \emph{C\link{dB}} which contains amplitudes of
at each ½ Quefrency (in ms).
}
\details{
The function is a re-write of the \link[wrassp:cepstrum]{wrassp::cepstrum} function, but
with media pre-conversion, better checking of preconditions such as the
input file existence, structured logging, and the use of a more modern
framework for user feedback. This version includes Rcpp optimizations
for improved performance on large batches of files.

The native file type of this function is "wav" files (in "pcm_s16le"
format), SUNs "au", NIST, or CSL formats (kay or NSP extension). Input
signal conversion, when needed, is done by
\href{https://ffmpeg.org/libavcodec.html}{libavcodec} and the excellent \link[av:encoding]{av::av_audio_convert}
wrapper function
}
\note{
This function takes some time to apply but also result in data in a relatively large matrix.
It is therefore not usually efficient to store intermediate results in a cache.
However, if the number of signals it will be applied to
is \emph{very} large, then caching of results may be warranted.
}
\examples{
# get path to audio file
path2wav <- list.files(system.file("samples","sustained", package = "superassp"), pattern = glob2rx("a1.wav"), full.names = TRUE)

# calulate cepstrum
res <- trk_cepstrum(path2wav, toFile=FALSE)

# plot cepstral values at midpoint of signal
plot(y=res[["C[dB]"]][400,],
    x=seq(1,ncol(res[["C[dB]"]])),
    type='l',
    xlab='Quefrency (ms)',
    ylab='Amplitude (dB)')

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{dftSpectrum}}, \code{\link{cssSpectrum}}, \code{\link{lpsSpectrum}};
all derived from libassp's spectrum function

\link[wrassp:cepstrum]{wrassp::cepstrum}

\link{AsspWindowTypes}

\link[av:encoding]{av::av_audio_convert}
}
\author{
Raphael Winkelmann

Lasse Bombien

Fredrik Nylén
}
