% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_python_yin.R
\name{trk_yin}
\alias{trk_yin}
\title{Pitch detection using the YIN metod}
\usage{
trk_yin(listOfFiles, ...)
}
\arguments{
\item{trough_threshold}{The absolute threshold for peak estimation.}

\item{center}{Should analysis windows be centered around the time of the
window (\code{TRUE}, the default) or should the window be considered to have
started at the indicated time point (\code{FALSE}).}

\item{pad_mode}{The mode in which padding occurs. Ignored if \code{center} is not
\code{TRUE}. Padding occurs in the python library librosa, and the user should
therefore consult the manual of the NumPy library function
\href{https://numpy.org/doc/stable/reference/generated/numpy.pad.html#numpy-pad}{numpy.pad} for other options.}
}
\value{
An SSFF track object containing two tracks (f0 and pitch) that are either returned (toFile == FALSE) or stored on disk.
}
\description{
This function applies the YIN
\insertCite{Cheveigné.2002.10.1121/1.1458024}{superassp} method to estimate
the fundamental frequency.
}
\details{
This function calls the librosa \insertCite{brian_mcfee_2022_6097378}{superassp} Python library to load the audio data an
make pitch related estimates.
}
\references{
\insertAllCited{}
}
\seealso{
pyin
}
