% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_track_core.R
\name{validate_json_track}
\alias{validate_json_track}
\title{Validate JsonTrackObj}
\usage{
validate_json_track(obj)
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Validate JsonTrackObj
}
