% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ssff.R
\name{write_ssff}
\alias{write_ssff}
\title{Write an AsspDataObj to an SSFF file}
\usage{
write_ssff(dobj, file = attr(dobj, "filePath"))
}
\arguments{
\item{dobj}{An \code{AsspDataObj}.}

\item{file}{Output file path. Defaults to the \code{filePath} attribute of \code{dobj}.}
}
\value{
Invisibly, \code{file}.
}
\description{
User-facing wrapper around the ASSP C-level writer.
Interface mirrors the legacy \code{write.AsspDataObj}.
}
\seealso{
\code{\link{write_json_track}} for JSTF (JSON) output.
}
