% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_ssff.R
\name{yin}
\alias{yin}
\title{Pitch detection using the YIN metod}
\usage{
yin(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  windowSize = 30,
  minF = 70,
  maxF = 200,
  trough_threshold = 0.1,
  center = TRUE,
  pad_mode = "constant",
  explicitExt = "yip",
  outputDirectory = NULL,
  toFile = TRUE
)
}
\arguments{
\item{listOfFiles}{A vector of file paths to wav files.}

\item{beginTime}{The start time of the section of the sound file that should be processed.}

\item{endTime}{The end time of the section of the sound file that should be processed.}

\item{windowShift}{The measurement interval (frame duration), in seconds.}

\item{minF}{Candidate f0 frequencies below this frequency will not be considered.}

\item{maxF}{Candidates above this frequency will be ignored.}

\item{trough_threshold}{The absolute threshold for peak estimation.}

\item{center}{Should analysis windows be centered around the time of the
window (\code{TRUE}, the default) or should the window be considered to have
started at the indicated time point (\code{FALSE}).}

\item{pad_mode}{The mode in which padding occurs. Ignored if \code{center} is not
\code{TRUE}. Padding occurs in the python library librosa, and the user should
therefore consult the manual of the NumPy library function
\href{https://numpy.org/doc/stable/reference/generated/numpy.pad.html#numpy-pad}{numpy.pad} for other options.}

\item{explicitExt}{the file extension that should be used.}

\item{outputDirectory}{set an explicit directory for where the signal file will be written. If not defined, the file will be written to the same directory as the sound file.}

\item{toFile}{write the output to a file? The file will be written in  \code{outputDirectory}, if defined, or in the same directory as the soundfile.}
}
\value{
An SSFF track object containing two tracks (f0 and pitch) that are either returned (toFile == FALSE) or stored on disk.
}
\description{
This function applies the YIN
\insertCite{Cheveigné.2002.10.1121/1.1458024}{superassp} method to estimate
the fundamental frequency.
}
\details{
This function calls the \insertCite{librosa
brian_mcfee_2022_6097378}{superassp} Python library to load the audio data an
make pitch related estimates.
}
\references{
\insertAllCited{}
}
